% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto-nbd-abe.R
\name{abe.GenerateData}
\alias{abe.GenerateData}
\title{Generate artificial data which follows Abe's Pareto/NBD model variant.}
\usage{
abe.GenerateData(n, T.cal, T.star, params, return.elog = FALSE)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{list of parameters: {beta, gamma}}

\item{return.elog}{if \code{TRUE} then the event-log is returned as well; decreases performance}
}
\value{
2-elemnt list
}
\description{
Returns 2-element list
* cbs: data.frame with 'cust', \code{x}, \code{t.x}, \code{T.cal}, 'T.star', 'x.star' 
       this is the summary statistics data.frame which contains all 
       needed information for parameter estimation
* elog: data.frame with 'cust', \code{t}
}
\examples{
# generate artificial Pareto/NBD Abe with 2 covariates
params <- list()
params$beta  <- matrix(c(0.18, -2.5, 0.5, -0.3, -0.2, 0.8), byrow = T, ncol = 2)
params$gamma <- matrix(c(0.05, 0.1, 0.1, 0.2), ncol = 2)
data <- abe.GenerateData(n = 5000, T.cal = 32, T.star = 32, params, return.elog = TRUE)
cbs <- data$cbs  # customer by sufficient summary statistic - one row per customer
elog <- data$elog  # Event log - one row per event/purchase
}

