% Generated by roxygen2 (4.0.1): do not edit by hand
\name{abe.mcmc.PAlive}
\alias{abe.mcmc.PAlive}
\title{Calculates P(alive) based on MCMC draws}
\usage{
abe.mcmc.PAlive(cal.cbs, draws)
}
\arguments{
\item{cal.cbs}{data.frame with column \code{T.cal}}

\item{draws}{MCMC draws returned by \code{\link{abe.mcmc.DrawParameters}}}
}
\value{
numeric vector of probabilities
}
\description{
Calculates P(alive) based on MCMC draws
}

