% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.R
\name{bgcnbd.cbs.LL}
\alias{bgcnbd.cbs.LL}
\title{Calculate the log-likelihood of the BG/CNBD-k model}
\usage{
bgcnbd.cbs.LL(params, cal.cbs, dropout_at_zero = FALSE)
}
\arguments{
\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r},
\code{alpha}, \code{a} and \code{b} in that order.}

\item{cal.cbs}{calibration period CBS. It must contain columns for frequency 
\code{x}, for recency \code{t.x}, for sum of logarithmic interpurchase
times \code{litt} and total time observed \code{T.cal}. Optionally a column
\code{custs} can be provided, which represents number of customers with a
specific combination of frequency \code{x} and \code{T.cal}.}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
the total log-likelihood for the provided data.
}
\description{
Calculate the log-likelihood of the BG/CNBD-k model
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

