% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.R
\name{bgcnbd.ConditionalExpectedTransactions}
\alias{bgcnbd.ConditionalExpectedTransactions}
\title{BG/CNBD-k Conditional Expected Transactions}
\usage{
bgcnbd.ConditionalExpectedTransactions(params, T.star, x, t.x, T.cal,
  dropout_at_zero = FALSE)
}
\arguments{
\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a}
and \code{b} in that order.}

\item{T.star}{length of time for which we are calculating the expected number
of transactions.}

\item{x}{number of repeat transactions in the calibration period T.cal, or a 
vector of calibration period frequencies.}

\item{t.x}{recency, i.e. length between first and last transaction during 
calibration period.}

\item{T.cal}{length of calibration period, or a vector of calibration period 
lengths.}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
Number of transactions a customer is expected to make in a time 
  period of length t, conditional on their past behavior. If any of the input
  parameters has a length greater than 1, this will be a vector of expected 
  number of transactions.
}
\description{
Uses BG/CNBD-k model parameters and a customer's past transaction behavior 
to return the number of transactions they are expected to make in a given 
time period.
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

