% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.R
\name{bgcnbd.EstimateParameters}
\alias{bgcnbd.EstimateParameters}
\title{Parameter Estimation for the BG/CNBD-k model}
\usage{
bgcnbd.EstimateParameters(cal.cbs, k = NULL, par.start = c(1, 3, 1, 3),
  max.param.value = 10000, trace = 0, dropout_at_zero = FALSE)
}
\arguments{
\item{cal.cbs}{calibration period CBS. It must contain columns for frequency 
\code{x}, for recency \code{t.x} and total time observed \code{T.cal}.
Optionally a column \code{custs} can be provided, which represents number
of customers with a specific combination of frequency \code{x}, recency
\code{t.x} and \code{T.cal}.}

\item{k}{specified degree of regularity for Erlang-k distributed 
interpurchase times; needs to be integer-value; if this is not specified, 
then grid search from 1 to 12 is performed; this however requires column 
\code{litt} to be present in cal.cbs, which represents sum of logarithmic 
interpurchase times during calibration period;}

\item{par.start}{initial BG/CNBD-k parameters - a vector with \code{r},
\code{alpha}, \code{a} and \code{b} in that order.}

\item{max.param.value}{the upper bound on parameters}

\item{trace}{print logging step every \code{trace} iteration}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
list of estimated parameters
}
\description{
Estimates parameters for the BG/CNBD-k via Maximum Likelihood Estimation.
}
\examples{
cbs <- cdnow.sample()$cbs # load CDNow summary data
params <- bgcnbd.EstimateParameters(cbs)
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}
\seealso{
\code{\link{elog2cbs}}
}

