% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.R
\name{bgcnbd.Expectation}
\alias{bgcnbd.Expectation}
\title{BG/CNBD-k Expectation}
\usage{
bgcnbd.Expectation(params, t, dropout_at_zero = FALSE)
}
\arguments{
\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{t}{length of time for which we are calculating the expected number of repeat transactions.}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
Number of repeat transactions a customer is expected to make in a time period of length t.
}
\description{
Returns the number of repeat transactions that a randomly chosen customer
(for whom we have no prior information) is expected to make in a given time
period.
}
\details{
E(X(t) | k, r, alpha, a, b)
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

