% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.R
\name{bgcnbd.ExpectedCumulativeTransactions}
\alias{bgcnbd.ExpectedCumulativeTransactions}
\title{BG/CNBD-k Expected Cumulative Transactions}
\usage{
bgcnbd.ExpectedCumulativeTransactions(params, T.cal, T.tot, n.periods.final,
  dropout_at_zero = FALSE)
}
\arguments{
\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{T.cal}{a vector to represent customers' calibration period lengths (in
other words, the 'T.cal' column from a customer-by-sufficient-statistic
matrix).}

\item{T.tot}{end of holdout period. Must be a single value, not a vector.}

\item{n.periods.final}{number of time periods in the calibration and holdout periods.}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
Vector of expected cumulative total repeat transactions by all customers.
}
\description{
Calculates the expected cumulative total repeat transactions by all customers
for the calibration and holdout periods.
}
\seealso{
\code{\link{bgnbd.ExpectedCumulativeTransactions}}
}

