% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.R
\name{bgcnbd.GenerateData}
\alias{bgcnbd.GenerateData}
\title{Simulate data according to BG/CNBD-k model assumptions}
\usage{
bgcnbd.GenerateData(n, T.cal, T.star = NULL, params, return.elog = FALSE,
  dropout_at_zero = FALSE)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period; if vector then it is assumed that
customers have different 'birth' dates, i.e. max(T.cal)-T.cal}

\item{T.star}{length(s) of holdout period(s); assumed to be same for all
customers}

\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r},
\code{alpha}, \code{a} and \code{b} in that order.}

\item{return.elog}{boolean - if \code{TRUE} then the event log is returned in 
addition to the CBS summary}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
list with elements \code{cbs} and \code{elog} containing data.frames
}
\description{
Simulate data according to BG/CNBD-k model assumptions
}
\examples{
params <- c(k = 3, r = 0.85, alpha = 1.45, a = 0.79, b = 2.42)
data <- bgcnbd.GenerateData(n = 4000, T.cal = 24, T.star = 32, params, return.elog = TRUE)
cbs <- data$cbs  # customer by sufficient summary statistic - one row per customer
elog <- data$elog  # Event log - one row per event/purchase
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

