% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.r
\name{bgcnbd.LL}
\alias{bgcnbd.LL}
\title{Calculate the log-likelihood of the BG/CNBD-k model}
\usage{
bgcnbd.LL(params, x, t.x, T.cal, litt, dropout_at_zero = FALSE)
}
\arguments{
\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a}
and \code{b} in that order.}

\item{x}{frequency, i.e. number of re-purchases}

\item{t.x}{recency, i.e. time elapsed from first purchase to last purchase}

\item{T.cal}{total time of observation period}

\item{litt}{sum of logarithmic interpurchase times}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
a vector of log-likelihoods
}
\description{
Calculate the log-likelihood of the BG/CNBD-k model
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

