% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.r
\name{bgcnbd.PlotFrequencyInCalibration}
\alias{bgcnbd.PlotFrequencyInCalibration}
\title{BG/CNBD-k Plot Frequency in Calibration Period}
\usage{
bgcnbd.PlotFrequencyInCalibration(params, cal.cbs, censor = NULL,
  plotZero = TRUE, xlab = "Calibration period transactions",
  ylab = "Customers", title = "Frequency of Repeat Transactions",
  dropout_at_zero = FALSE)
}
\arguments{
\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{cal.cbs}{calibration period CBS (customer by sufficient statistic). It
must contain columns for frequency ("x") and total time observed ("T.cal").}

\item{censor}{integer used to censor the data.}

\item{plotZero}{If FALSE, the histogram will exclude the zero bin.}

\item{xlab}{descriptive label for the x axis.}

\item{ylab}{descriptive label for the y axis.}

\item{title}{title placed on the top-center of the plot.}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
Calibration period repeat transaction frequency comparison matrix (actual vs. expected).
}
\description{
Plots a histogram and returns a matrix comparing the actual and expected
number of customers who made a certain number of repeat transactions in the
calibration period, binned according to calibration period frequencies.
}
\examples{
set.seed(1)
params <- c(k=3, r=0.85, alpha=1.45, a=0.79, b=2.42)
cal.cbs <- bgcnbd.GenerateData(n=5000, 32, 0, params)$cbs
bgcnbd.PlotFrequencyInCalibration(params, cal.cbs, censor = 7)
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

