% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.r
\name{bgcnbd.PlotTrackingInc}
\alias{bgcnbd.PlotTrackingInc}
\title{BG/CNBD-k Tracking Incremental Transactions Comparison}
\usage{
bgcnbd.PlotTrackingInc(params, T.cal, T.tot, actual.inc.tracking.data,
  xlab = "Week", ylab = "Transactions", xticklab = NULL,
  title = "Tracking Weekly Transactions", ymax = NULL,
  dropout_at_zero = FALSE)
}
\arguments{
\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{T.cal}{a vector to represent customers' calibration period lengths (in
other words, the "T.cal" column from a customer-by-sufficient-statistic
matrix).}

\item{T.tot}{end of holdout period. Must be a single value, not a vector.}

\item{xlab}{descriptive label for the x axis.}

\item{ylab}{descriptive label for the y axis.}

\item{xticklab}{vector containing a label for each tick mark on the x axis.}

\item{title}{title placed on the top-center of the plot.}

\item{ymax}{upper boundary for y axis.}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}

\item{actual.cu.tracking.data}{vector containing the cumulative number of
repeat transactions made by customers for each period in the total time
period (both calibration and holdout periods).}
}
\value{
Matrix containing actual and expected incremental repeat transactions.
}
\description{
Plots the actual and expected incremental total repeat transactions by all
customers for the calibration and holdout periods, and returns this
comparison in a matrix.
}
\seealso{
\code{\link{bgnbd.PlotTrackingInc}}
}

