% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.R
\name{bgcnbd.pmf}
\alias{bgcnbd.pmf}
\title{BG/CNBD-k Probability Mass Function}
\usage{
bgcnbd.pmf(params, t, x, dropout_at_zero = FALSE)
}
\arguments{
\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{t}{length of time for which we are calculating the expected number of 
transactions.}

\item{x}{number of transactions for which probability is calculated.}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
P(X(t)=x|r,alpha,a,b). If any of the input parameters has a length
  greater than 1, this will be a vector of expected number of transactions.
}
\description{
Uses BG/CNBD-k model parameters to return the probability distribution of
purchase frequencies for a random customer in a given time period, i.e.
P(X(t)=x|r,alpha,a,b)
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

