% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bgnbd.cbs.LL}
\alias{bgnbd.cbs.LL}
\title{Calculate the log-likelihood of the BG/NBD model}
\usage{
bgnbd.cbs.LL(params, cal.cbs)
}
\arguments{
\item{params}{BG/NBD parameters - a vector with \code{r}, \code{alpha}, \code{a} and
\code{b} in that order.}

\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
\code{x}, for recency \code{t.x} and total time observed \code{T.cal}. Optionally a
column \code{custs} can be provided, which represents number of customers with a
specific combination of frequency \code{x} and \code{T.cal}.}
}
\value{
the total log-likelihood for the provided data.
}
\description{
Calculate the log-likelihood of the BG/NBD model
}
\seealso{
\code{\link{bgnbd.EstimateParameters}}
}

