% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bgnbd.ConditionalExpectedTransactions}
\alias{bgnbd.ConditionalExpectedTransactions}
\title{BG/NBD Conditional Expected Transactions}
\usage{
bgnbd.ConditionalExpectedTransactions(params, T.star, x, t.x, T.cal)
}
\arguments{
\item{params}{BG/NBD parameters - a vector with 'r', 'alpha', 'a' and 'b' in
that order.}

\item{T.star}{length of time for which we are calculating the expected number
of transactions.}

\item{x}{number of repeat transactions in the calibration period T.cal, or a
vector of calibration period frequencies.}

\item{t.x}{recency, i.e. length between first and last transaction during
calibration period.}

\item{T.cal}{length of calibration period, or a vector of calibration period
lengths.}
}
\value{
Number of transactions a customer is expected to make in a time
  period of length t, conditional on their past behavior. If any of the input
  parameters has a length greater than 1, this will be a vector of expected
  number of transactions.
}
\description{
Uses BG/NBD model parameters and a customer's past transaction behavior to
return the number of transactions they are expected to make in a given time
period.
}
\examples{

set.seed(1)

# generate artificial BG/NBD data 
n      <- 1000 # no. of customers
T.cal  <- 32   # length of calibration period
T.star <- 32   # length of hold-out period
params <- list(r=0.85, alpha=4.45, # purchase frequency lambda_i ~ Gamma(r, alpha)
               a=0.79, b=2.42)     # dropout probability p_i ~ Beta(a, b)

cbs <- bgnbd.GenerateData(n, T.cal, T.star, params)$cbs

# estimate parameters, and compare to true parameters
est <- bgnbd.EstimateParameters(cbs[, c("x", "t.x", "T.cal")])
rbind(params, est=round(est, 2))
#        r    alpha a    b   
# params 0.85 4.45  0.79 2.42
# est    0.82 4.32  0.62 1.98
# -> underlying parameters are successfully identified via Maximum Likelihood Estimation

# estimate future transactions in holdout-period
cbs$x.est <- bgnbd.ConditionalExpectedTransactions(est, cbs$T.star, cbs$x, cbs$t.x, cbs$T.cal)

# compare forecast accuracy to naive forecast
c("bgnbd"=sqrt(mean((cbs$x.star-cbs$x.est)^2)),
  "naive"=sqrt(mean((cbs$x.star-cbs$x)^2)))
#    bgnbd    naive 
# 2.386392 3.688767
# -> BG/NBD forecast better than naive forecast

# estimate P(alive)
cbs$palive <- bgnbd.PAlive(est, cbs$x, cbs$t.x, cbs$T.cal)

# compare to true (usually unobserved) alive status
prop.table(table(cbs$palive>.5, cbs$alive))
#       FALSE  TRUE
# FALSE 0.375 0.065
# TRUE  0.176 0.384
# -> 76\% of customers are correctly classified
}
\seealso{
bgnbd.EstimateParameters
}

