% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bgnbd.EstimateParameters}
\alias{bgnbd.EstimateParameters}
\title{Parameter Estimation for the BG/NBD model}
\usage{
bgnbd.EstimateParameters(cal.cbs, par.start = c(1, 1, 1, 1),
  max.param.value = 10000)
}
\arguments{
\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
'x', for recency 't.x.' and total time observed 'T.cal'. Optionally a
column 'custs' can be provided, which represents number of customers with a
specific combination of frequency 'x', recency 't.x' and 'T.cal'.}

\item{par.start}{initial BG/NBD parameters - a vector with 'r', 'alpha', 'a' and
'b' in that order.}

\item{max.param.value}{the upper bound on parameters}
}
\value{
list of estimated parameters
}
\description{
Estimates parameters for the BG/NBD model via Maximum Likelihood Estimation.
}
\examples{
# generate artificial data
params <- list(r=0.24, alpha=4.41, a=0.79, b=2.42)
cbs <- bgnbd.GenerateData(n=1000, T.cal=32, T.star=32, params=params)$cbs

# estimate parameters, and compare to true parameters
est <- bgnbd.EstimateParameters(cbs)
rbind(params, est=round(est, 3))
#        r     alpha a     b
# params 0.24  4.41  0.79  2.42
# est    0.238 4.408 0.522 1.288

# estimate future transactions, and compare to actuals from holdout period
cbs$x.est <- bgnbd.ConditionalExpectedTransactions(est, cbs$T.star, cbs$x, cbs$t.x, cbs$T.cal)
sqrt(mean((cbs$x.star-cbs$x.est)^2))
# 1.34969
}
\references{
Fader, Peter S., Bruce GS Hardie, and Ka Lok Lee. "Counting Your Customers the Easy Way: An Alternative to the Pareto/NBD Model." Marketing Science 24.2 (2005): 275-284.
}

