% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bgnbd.GenerateData}
\alias{bgnbd.GenerateData}
\title{Simulate data according to BG/NBD model assumptions}
\usage{
bgnbd.GenerateData(n, T.cal, T.star, params, return.elog = F)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{BG/NBD parameters - a list with 'r', 'alpha', 'a' and 'b'}

\item{return.elog}{boolean - if TRUE then the event log is returned in
addition to the CBS summary}
}
\value{
list with elements 'cbs' and 'elog' containing data.frames
}
\description{
Simulate data according to BG/NBD model assumptions
}

