% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bgnbd.LL}
\alias{bgnbd.LL}
\title{Calculate the log-likelihood of the BG/NBD model}
\usage{
bgnbd.LL(params, x, t.x, T.cal)
}
\arguments{
\item{params}{BG/NBD parameters - a vector with 'r', 'alpha', 'a' and
'b' in that order.}

\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
'x', for recency 't.x.' and total time observed 'T.cal'. Optionally a
column 'custs' can be provided, which represents number of customers with a
specific combination of frequency 'x' and 'T.cal'.}
}
\value{
a vector of log-likelihoods
}
\description{
Calculate the log-likelihood of the BG/NBD model
}

