% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cbgcnbd.cbs.LL}
\alias{cbgcnbd.cbs.LL}
\title{Calculate the log-likelihood of the CBG/CNBD-k model}
\usage{
cbgcnbd.cbs.LL(params, k, cal.cbs)
}
\arguments{
\item{params}{CBG/CNBD-k parameters - a vector with 'r', 'alpha', 'a' and 'b'
in that order.}

\item{k}{specified degree of regularity for Erlang-k distributed
interpurchase times}

\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
'x', for recency 't.x.' and total time observed 'T.cal'. Optionally a
column 'custs' can be provided, which represents number of customers with a
specific combination of frequency 'x' and 'T.cal'.}
}
\value{
the total log-likelihood for the provided data.
}
\description{
Calculate the log-likelihood of the CBG/CNBD-k model
}

