% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cbgcnbd.EstimateParameters}
\alias{cbgcnbd.EstimateParameters}
\title{Parameter Estimation for the CBG/CNBD-k model}
\usage{
cbgcnbd.EstimateParameters(cal.cbs, k, par.start = c(1, k, 1, 1),
  max.param.value = 10000)
}
\arguments{
\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
'x', for recency 't.x.' and total time observed 'T.cal'. Optionally a
column 'custs' can be provided, which represents number of customers with a
specific combination of frequency 'x', recency 't.x' and 'T.cal'.}

\item{k}{specified degree of regularity for Erlang-k distributed
interpurchase times}

\item{par.start}{initial CBG/CNBD-k parameters - a vector with 'r', 'alpha',
'a' and 'b' in that order.}

\item{max.param.value}{the upper bound on parameters}
}
\value{
list of estimated parameters
}
\description{
Estimates parameters for the CBG/CNBD-k model via Maximum Likelihood
Estimation.
}
\examples{
# generate artificial data
k <- 3
params <- list(r=0.53, alpha=1.25, a=1.29, b=1.85)
data <- cbgcnbd.GenerateData(n=1000, k=k, T.cal=32, T.star=32, params=params, return.elog=T)
cbs <- data$cbs
elog <- data$elog

# estimate parameters, and compare to true parameters
k.est <- cbgcnbd.EstimateRegularity(elog)
# 3.175759; -> CBG/CNBD-3 should be modelled
est <- cbgcnbd.EstimateParameters(cbs, round(k.est))
rbind(params, est=round(est, 3))
#        r     alpha a     b
# params 0.53  1.25  1.29  1.85
# est    0.443 1.217 1.324 2.695

# estimate future transactions, and compare to actuals from holdout period
cbs$x.est <- cbgcnbd.ConditionalExpectedTransactions(est, k=round(k.est), cbs$T.star, cbs$x, cbs$t.x, cbs$T.cal)
sqrt(mean((cbs$x.star-cbs$x.est)^2))
# 1.168548
}
\references{
Platzer, Michael. "Stochastic models of noncontractual consumer
  relationships." Master of Science in Business Administration thesis, Vienna
  University of Economics and Business Administration, Austria (2008).
  \url{https://sites.google.com/site/michaelplatzer/stochastic-models-of-noncontractual-consumer-relationships}
}

