% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cbgcnbd.EstimateRegularity}
\alias{cbgcnbd.EstimateRegularity}
\title{Estimate Regularity in Intertransaction Timings}
\usage{
cbgcnbd.EstimateRegularity(elog, method = "wheat", plot = F)
}
\arguments{
\item{elog}{data.frame with transaction logs; requires columns customer-id
'cust' and transaction time 't' or 'date'}

\item{method}{'wheat', 'mle', 'mle-minka', 'mle-thom', 'cv'}

\item{plot}{if TRUE then distribution of estimated regularity will be plotted}
}
\value{
estimated real-valued regularity parameter; rounded to an integer,
  this can be used as 'k' for estimating CBG/CNBD-k models
}
\description{
Estimates degree of regularity of intertransaction timings of a customer cohort,
}
\details{
This is done
 - by either assuming same regularity across all customers; this then only
     require three transactions per customer
     method: wheat
 - or by estimating regularity for each customer seperately (as the shape
     parameter of a fitted gamma distribution), and then return the
     median across estimates; this requires min. 10 transactions per customer
     methods: mle, mle-minka, mle-thom, cv
}
\seealso{
cbgcnbd.EstimateParameters
}

