% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cbgcnbd.GenerateData}
\alias{cbgcnbd.GenerateData}
\title{Simulate data according to CBG/CNBD-k model assumptions}
\usage{
cbgcnbd.GenerateData(n, k, T.cal, T.star = T.cal, params, return.elog = F)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{CBG/CNBD-k parameters - a list with 'r', 'alpha', 'a' and 'b'}

\item{return.elog}{boolean - if TRUE then the event log is returned in
addition to the CBS summary}
}
\value{
list with elements 'cbs' and 'elog' containing data.frames
}
\description{
Simulate data according to CBG/CNBD-k model assumptions
}

