% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cbgcnbd.PAlive}
\alias{cbgcnbd.PAlive}
\title{CBG/CNBD-k P(alive)}
\usage{
cbgcnbd.PAlive(params, k, x, t.x, T.cal)
}
\arguments{
\item{params}{CBG/CNBD-k parameters - a vector with 'r', 'alpha', 'a' and 'b'
in that order.}

\item{k}{specified degree of regularity for Erlang-k distributed
interpurchase times}

\item{x}{number of repeat transactions in the calibration period T.cal, or a
vector of calibration period frequencies.}

\item{t.x}{recency, i.e. length between first and last transaction during
calibration period.}

\item{T.cal}{length of calibration period, or a vector of calibration period
lengths.}
}
\value{
Probability that the customer is still alive at the end of the
  calibration period.
}
\description{
Uses CBG/CNBD-k model parameters and a customer's past transaction behavior
to return the probability that they are still alive at the end of the
calibration period.
}

