% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cbgcnbd.Px}
\alias{cbgcnbd.Px}
\title{CBD/CNBD-k Unconditional Probability Distribution of Transactions}
\usage{
cbgcnbd.Px(params, t, x)
}
\arguments{
\item{params}{CBG/CNBD-k parameters - a vector with 'k', \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{t}{length of time for which we are calculating the expected number of
transactions.}

\item{x}{number of transactions for which probability is calculated.}
}
\value{
P(X(t)=x|r,alpha,a,b). If any of the input parameters has a length
  greater than 1, this will be a vector of expected number of transactions.
}
\description{
Uses CBG/CNBD-k model parameters to return the probability distribution of
purchase frequencies for a random customer in a given time period, i.e.
P(X(t)=x|r,alpha,a,b)
}
\seealso{
\code{\link{cbgcnbd.EstimateParameters}}
}

