% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cbgcnbd.UnconditionalExpectedTransactions}
\alias{cbgcnbd.UnconditionalExpectedTransactions}
\title{CBD/CNBD-k Unconditional Expected Transactions}
\usage{
cbgcnbd.UnconditionalExpectedTransactions(params, t)
}
\arguments{
\item{params}{CBG/CNBD-k parameters - a vector with 'k', \code{r}, \code{alpha}, \code{a}
and \code{b} in that order.}

\item{t}{length of time for which we are calculating the expected number of
transactions.}
}
\value{
Number of transactions a customer is expected to make in a time
  period of length t. If any of the input parameters has a length greater
  than 1, this will be a vector of expected number of transactions.
}
\description{
Uses CBG/CNBD-k model parameters to return the number of transactions a
random customer is expected to make in a given time period.
}
\seealso{
\code{\link{cbgcnbd.EstimateParameters}}
}

