% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cbgnbd.cbs.LL}
\alias{cbgnbd.cbs.LL}
\title{Calculates the log-likelihood of the CBG/NBD model}
\usage{
cbgnbd.cbs.LL(params, cal.cbs)
}
\description{
Calculates the log-likelihood of the CBG/NBD model
}
\seealso{
\code{\link{cbgcnbd.cbs.LL}}
}

