% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cbgnbd.ConditionalExpectedTransactions}
\alias{cbgnbd.ConditionalExpectedTransactions}
\title{CBG/NBD Conditional Expected Transactions}
\usage{
cbgnbd.ConditionalExpectedTransactions(params, T.star, x, t.x, T.cal)
}
\description{
CBG/NBD Conditional Expected Transactions
}
\seealso{
\code{\link{cbgcnbd.ConditionalExpectedTransactions}}
}

