% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cbgnbd.EstimateParameters}
\alias{cbgnbd.EstimateParameters}
\title{Parameter Estimation for the CBG/NBD model}
\usage{
cbgnbd.EstimateParameters(cal.cbs, par.start = c(1, 1, 1, 1),
  max.param.value = 10000)
}
\description{
Parameter Estimation for the CBG/NBD model
}
\seealso{
\code{\link{cbgcnbd.EstimateParameters}}
}

