% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cdnow.sample}
\alias{cdnow.sample}
\title{CDNow Sample Data}
\usage{
cdnow.sample()
}
\value{
list with two data.frames: `elog` for the event logs and `cbs` for
  the customer by sufficient summary
}
\description{
This is a convenience wrapper for data('cdnowElog', package='BTYD'), with
same-day transactions being merged together for each customer.
}
\references{
Fader, Peter S. and Bruce G.,S. Hardie, (2001), "Forecasting
  Repeat Sales at CDNOW: A Case Study," Interfaces, 31 (May-June), Part 2 of
  2, S94-S107.
}
\seealso{
\code{\link{elog2cbs}}
}

