% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dc.PlotFreqVsConditionalExpectedFrequency}
\alias{dc.PlotFreqVsConditionalExpectedFrequency}
\title{Generic Method for Plotting Frequency vs. Conditional Expected Frequency}
\usage{
dc.PlotFreqVsConditionalExpectedFrequency(
  x,
  actual,
  expected,
  censor,
  xlab,
  ylab,
  xticklab,
  title
)
}
\description{
Generic Method for Plotting Frequency vs. Conditional Expected Frequency
}
\keyword{internal}
