% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dc.PlotRecVsConditionalExpectedFrequency}
\alias{dc.PlotRecVsConditionalExpectedFrequency}
\title{Generic Method for Plotting Frequency vs. Conditional Expected Frequency}
\usage{
dc.PlotRecVsConditionalExpectedFrequency(
  t.x,
  actual,
  expected,
  xlab,
  ylab,
  xticklab,
  title
)
}
\description{
Generic Method for Plotting Frequency vs. Conditional Expected Frequency
}
\keyword{internal}
