% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dc.PlotTracking}
\alias{dc.PlotTracking}
\title{Generic Method for Tracking Plots}
\usage{
dc.PlotTracking(actual, expected, T.cal = NULL, xlab = "", ylab = "",
  title = "", xticklab = NULL, ymax = NULL)
}
\arguments{
\item{actual}{actual}

\item{expected}{expected}

\item{T.cal}{T.cal}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{title}{title}

\item{xticklab}{xticklab}

\item{ymax}{ymax}
}
\value{
Matrix
}
\description{
Generic Method for Tracking Plots
}
\seealso{
\code{\link{mcmc.PlotTrackingCum}} \code{\link{mcmc.PlotTrackingInc}} \code{\link{bgcnbd.PlotTrackingCum}} \code{\link{bgcnbd.PlotTrackingInc}}
}

