% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{elog2cum}
\alias{elog2cum}
\title{Aggregate Event Log to cumulative number of repeat transactions}
\usage{
elog2cum(elog, by = 7)
}
\arguments{
\item{elog}{Event Log}

\item{by}{defaults to 7, which means weekly numbers}
}
\value{
numeric vector
}
\description{
Aggregate Event Log to cumulative number of repeat transactions
}
\examples{
elog <- cdnow.sample()$elog
cum <- elog2cum(elog)
plot(cum, typ="l")
}
\seealso{
\code{\link{elog2inc}}
}

