% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{elog2inc}
\alias{elog2inc}
\title{Aggregate Event Log to incremental number of (repeat) transactions}
\usage{
elog2inc(elog, by = 7, first = FALSE)
}
\arguments{
\item{elog}{Event log, a \code{data.frame} with columns \code{cust} and
transaction time \code{t} or \code{date}.}

\item{by}{Only return every \code{}-th number. Defaults to 7, and thus
returns weekly numbers.}

\item{first}{If TRUE, then the first transaction for each customer is being
counted as well}
}
\value{
Numeric vector of repeat transactions.
}
\description{
Aggregate Event Log to incremental number of (repeat) transactions
}
\examples{
data("groceryElog")
inc <- elog2inc(groceryElog)
plot(inc, typ="l", frame = FALSE)
}
\seealso{
\code{\link{elog2cum}}
}

