% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estimateRegularity}
\alias{estimateRegularity}
\title{Estimate Regularity in Intertransaction Timings}
\usage{
estimateRegularity(elog, method = "wheat", plot = FALSE, title = "")
}
\arguments{
\item{elog}{Event log, a \code{data.frame} with columns \code{cust} and
transaction time \code{t} or \code{date}}

\item{method}{Either \code{wheat}, \code{mle}, \code{mle-minka}, \code{mle-thom} or
\code{cv}.}

\item{plot}{If \code{TRUE} then distribution of estimated regularity will be
plotted.}

\item{title}{Plot title.}
}
\value{
Estimated real-valued regularity parameter.
}
\description{
The models (M)BG/CNBD-k and Pareto/GGG are capable of leveraging regularity
within transaction timings for improving forecast accuracy. This method
provides a quick check for the degree of regularity in the event timings. A
return value of close to 1 supports the assumption of exponentially
distributed intertransaction times, whereas values significantly larger than
1 reveal the presence of regularity. Estimation is either done by 1) assuming
a same degree of regularity across all customers (\code{method = "wheat"}), or 2) by
estimating regularity for each customer seperately, as the shape parameter of
a fitted gamma distribution, and then return the median across estimates. The
latter methods, though, require sufficient (>=10) transactions per customer.
}
\examples{
data("groceryElog")
estimateRegularity(groceryElog, plot = TRUE, method = 'wheat')
estimateRegularity(groceryElog, plot = TRUE, method = 'mle-minka')
estimateRegularity(groceryElog, plot = TRUE, method = 'mle-thom')
estimateRegularity(groceryElog, plot = TRUE, method = 'cv')
}
\references{
Wheat, Rita D., and Donald G. Morrison.  'Estimating purchase
  regularity with two interpurchase times.'

Dunn, Richard, Steven Reader, and Neil Wrigley. 'An investigation
  of the assumptions of the NBD model' Applied Statistics (1983): 249-259.

Wu, Couchen, and H-L. Chen. 'A consumer purchasing model with
  learning and departure behaviour.'  Journal of the Operational Research
  Society (2000): 583-591.

\url{http://research.microsoft.com/en-us/um/people/minka/papers/minka-gamma.pdf}
}

