% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggnbd.cbs.LL}
\alias{ggnbd.cbs.LL}
\title{Calculate the log-likelihood of the Gamma/Gompertz/NBD model}
\usage{
ggnbd.cbs.LL(params, cal.cbs)
}
\arguments{
\item{params}{Gamma/Gompertz/NBD parameters - a vector with 'r', 'alpha',
'b', 's' and 'beta' in that order.}

\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
'x', for recency 't.x.' and total time observed 'T.cal'. Optionally a
column 'custs' can be provided, which represents number of customers with a
specific combination of frequency 'x' and 'T.cal'.}
}
\value{
the total log-likelihood for the provided data.
}
\description{
Calculate the log-likelihood of the Gamma/Gompertz/NBD model
}
\seealso{
ggnbd.EstimateParameters
}

