% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-nbd.r
\name{ggnbd.ConditionalExpectedTransactions}
\alias{ggnbd.ConditionalExpectedTransactions}
\title{Gamma/Gompertz/NBD Conditional Expected Transactions}
\usage{
ggnbd.ConditionalExpectedTransactions(params, T.star, x, t.x, T.cal)
}
\arguments{
\item{params}{Gamma/Gompertz/NBD parameters - a vector with \code{r}, \code{alpha},
\code{b}, \code{s} and \code{beta} in that order.}

\item{T.star}{length of time for which we are calculating the expected number
of transactions.}

\item{x}{number of repeat transactions in the calibration period T.cal, or a 
vector of calibration period frequencies.}

\item{t.x}{recency, i.e. length between first and last transaction during 
calibration period.}

\item{T.cal}{length of calibration period, or a vector of calibration period 
lengths.}
}
\value{
Number of transactions a customer is expected to make in a time 
  period of length t, conditional on their past behavior. If any of the input
  parameters has a length greater than 1, this will be a vector of expected 
  number of transactions.
}
\description{
Uses Gamma/Gompertz/NBD model parameters and a customer's past transaction
behavior to return the number of transactions they are expected to make in a
given time period.
}
\seealso{
\code{\link{ggnbd.EstimateParameters}}
}

