% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-nbd.R
\name{ggnbd.EstimateParameters}
\alias{ggnbd.EstimateParameters}
\title{Parameter Estimation for Gamma/Gompertz/NBD model}
\usage{
ggnbd.EstimateParameters(cal.cbs, par.start = c(1, 1, 0.5, 1, 1),
  min.param.value = 1e-05, max.param.value = 10000, trace = 0)
}
\arguments{
\item{cal.cbs}{calibration period CBS. It must contain columns for frequency 
\code{x}, for recency \code{t.x} and total time observed \code{T.cal}. Optionally a 
column \code{custs} can be provided, which represents number of customers with a
specific combination of frequency \code{x}, recency \code{t.x} and \code{T.cal}.}

\item{par.start}{initial Gamma/Gompertz/NBD parameters - a vector with \code{r}, 
\code{alpha}, \code{b}, \code{s} and \code{beta} in that order.}

\item{min.param.value}{the lower bound on parameters}

\item{max.param.value}{the upper bound on parameters}

\item{trace}{print logging step every \code{trace} iteration}
}
\value{
list of estimated parameters
}
\description{
Estimates parameters for the Gamma/Gompertz/NBD model via Maximum Likelihood
Estimation.
}
\references{
Bemmaor, Albert C., and Nicolas Glady. 'Modeling Purchasing 
  Behavior with Sudden 'Death': A Flexible Customer Lifetime Model.'
  Management Science 58.5 (2012): 1012-1021.
}

