% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-nbd.r
\name{ggnbd.GenerateData}
\alias{ggnbd.GenerateData}
\title{Simulate data according to Gamma/Gompertz/NBD model assumptions}
\usage{
ggnbd.GenerateData(n, T.cal, T.star, params, return.elog = FALSE)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{Gamma/Gompertz/NBD parameters - a vector with \code{r}, \code{alpha},
\code{b}, \code{s} and \code{beta} in that order}

\item{return.elog}{boolean - if \code{TRUE} then the event log is returned in 
addition to the CBS summary}
}
\value{
list with elements \code{cbs} and \code{elog} containing data.frames
}
\description{
Simulate data according to Gamma/Gompertz/NBD model assumptions
}
\seealso{
\code{\link{ggnbd.EstimateParameters}}
}

