% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-nbd.R
\name{ggnbd.LL}
\alias{ggnbd.LL}
\title{Calculate the log-likelihood of the Gamma/Gompertz/NBD model}
\usage{
ggnbd.LL(params, x, t.x, T.cal)
}
\arguments{
\item{params}{Gamma/Gompertz/NBD parameters - a vector with \code{r}, \code{alpha}, 
\code{b}, \code{s} and \code{beta} in that order.}

\item{x}{frequency, i.e. number of re-purchases}

\item{t.x}{recency, i.e. time elapsed from first purchase to last purchase}

\item{T.cal}{total time of observation period}
}
\value{
a vector of log-likelihoods
}
\description{
Calculate the log-likelihood of the Gamma/Gompertz/NBD model
}
\seealso{
\code{\link{ggnbd.EstimateParameters}}
}

