% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-nbd.R
\name{ggnbd.PAlive}
\alias{ggnbd.PAlive}
\title{Gamma/Gompertz/NBD P(alive)}
\usage{
ggnbd.PAlive(params, x, t.x, T.cal)
}
\arguments{
\item{params}{Gamma/Gompertz/NBD parameters - a vector with \code{r}, \code{alpha},
\code{b}, \code{s} and \code{beta} in that order.}

\item{x}{number of repeat transactions in the calibration period T.cal, or a 
vector of calibration period frequencies.}

\item{t.x}{recency, i.e. length between first and last transaction during 
calibration period.}

\item{T.cal}{length of calibration period, or a vector of calibration period 
lengths.}
}
\value{
Probability that the customer is still alive at the end of the 
  calibration period.
}
\description{
Uses Gamma/Gompertz/NBD model parameters and a customer's past transaction
behavior to return the probability that they are still alive at the end of
the calibration period.
}
\seealso{
\code{\link{ggnbd.EstimateParameters}}
}

