% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.cbs.LL}
\alias{mbgcnbd.cbs.LL}
\title{Calculate the log-likelihood of the MBG/CNBD-k model}
\usage{
mbgcnbd.cbs.LL(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.cbs.LL}}}
}
\description{
Calculate the log-likelihood of the MBG/CNBD-k model
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

