% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.ConditionalExpectedTransactions}
\alias{mbgcnbd.ConditionalExpectedTransactions}
\title{MBG/CNBD-k Conditional Expected Transactions}
\usage{
mbgcnbd.ConditionalExpectedTransactions(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.ConditionalExpectedTransactions}}}
}
\description{
MBG/CNBD-k Conditional Expected Transactions
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

