% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.EstimateParameters}
\alias{mbgcnbd.EstimateParameters}
\title{Parameter Estimation for the MBG/CNBD-k model}
\usage{
mbgcnbd.EstimateParameters(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.EstimateParameters}}}
}
\description{
Parameter Estimation for the MBG/CNBD-k model
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

