% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.Expectation}
\alias{mbgcnbd.Expectation}
\title{MBG/CNBD-k Expectation}
\usage{
mbgcnbd.Expectation(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.Expectation}}}
}
\description{
MBG/CNBD-k Expectation
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

