% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.r
\name{mbgcnbd.ExpectedCumulativeTransactions}
\alias{mbgcnbd.ExpectedCumulativeTransactions}
\title{MBG/CNBD-k Expected Cumulative Transactions}
\usage{
mbgcnbd.ExpectedCumulativeTransactions(...)
}
\arguments{
\item{...}{Arguments to be passed to bgnbd.ExpectedCumulativeTransactions}
}
\description{
MBG/CNBD-k Expected Cumulative Transactions
}
\seealso{
\code{\link{bgcnbd.ExpectedCumulativeTransactions}}
}

