% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.GenerateData}
\alias{mbgcnbd.GenerateData}
\title{Simulate data according to MBG/CNBD-k model assumptions}
\usage{
mbgcnbd.GenerateData(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.GenerateData}}}
}
\description{
Simulate data according to MBG/CNBD-k model assumptions
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

