% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.PAlive}
\alias{mbgcnbd.PAlive}
\title{MBG/CNBD-k P(alive)}
\usage{
mbgcnbd.PAlive(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.PAlive}}}
}
\description{
MBG/CNBD-k P(alive)
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

