% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.PlotFrequencyInCalibration}
\alias{mbgcnbd.PlotFrequencyInCalibration}
\title{MBG/CNBD-k Plot Frequency in Calibration Period}
\usage{
mbgcnbd.PlotFrequencyInCalibration(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.PlotFrequencyInCalibration}}}
}
\description{
MBG/CNBD-k Plot Frequency in Calibration Period
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

