% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.PlotTrackingCum}
\alias{mbgcnbd.PlotTrackingCum}
\title{MBG/CNBD-k Tracking Cumulative Transactions Plot}
\usage{
mbgcnbd.PlotTrackingCum(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.PlotTrackingCum}}}
}
\description{
MBG/CNBD-k Tracking Cumulative Transactions Plot
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

