% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.PlotTrackingInc}
\alias{mbgcnbd.PlotTrackingInc}
\title{MBG/CNBD-k Tracking Incremental Transactions Comparison}
\usage{
mbgcnbd.PlotTrackingInc(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.PlotTrackingInc}}}
}
\description{
MBG/CNBD-k Tracking Incremental Transactions Comparison
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

