% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.pmf}
\alias{mbgcnbd.pmf}
\title{MBG/CNBD-k Unconditional Probability Distribution of Transactions}
\usage{
mbgcnbd.pmf(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgcnbd.pmf}}}
}
\description{
MBG/CNBD-k Unconditional Probability Distribution of Transactions
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

