% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mbg-nbd.r
\name{mbgnbd.cbs.LL}
\alias{mbgnbd.cbs.LL}
\title{Calculates the log-likelihood of the MBG/NBD model}
\usage{
mbgnbd.cbs.LL(params, cal.cbs)
}
\description{
Calculates the log-likelihood of the MBG/NBD model
}
\seealso{
\code{\link{mbgcnbd.cbs.LL}}
}

