% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.r
\name{mbgnbd.ConditionalExpectedTransactions}
\alias{mbgnbd.ConditionalExpectedTransactions}
\title{MBG/NBD Conditional Expected Transactions}
\usage{
mbgnbd.ConditionalExpectedTransactions(params, T.star, x, t.x, T.cal)
}
\description{
MBG/NBD Conditional Expected Transactions
}
\seealso{
\code{\link{mbgcnbd.ConditionalExpectedTransactions}}
}

