% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.r
\name{mbgnbd.EstimateParameters}
\alias{mbgnbd.EstimateParameters}
\title{Parameter Estimation for the MBG/NBD model}
\usage{
mbgnbd.EstimateParameters(cal.cbs, par.start = c(1, 1, 1, 1),
  max.param.value = 10000)
}
\description{
Parameter Estimation for the MBG/NBD model
}
\seealso{
\code{\link{mbgcnbd.EstimateParameters}}
}

