% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.R
\name{mbgnbd.Expectation}
\alias{mbgnbd.Expectation}
\title{MBG/NBD Expectation}
\usage{
mbgnbd.Expectation(params, t)
}
\arguments{
\item{params}{MBG/NBD parameters - a vector with \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{t}{length of time for which we are calculating the expected number of repeat transactions.}
}
\value{
Number of repeat transactions a customer is expected to make in a time period of length t.
}
\description{
Returns the number of repeat transactions that a randomly chosen customer
(for whom we have no prior information) is expected to make in a given time
period.
}
\details{
E(X(t) | k, r, alpha, a, b)
}
\references{
Batislam, E.P., M. Denizel, A. Filiztekin. 2007. Empirical
  validation and comparison of models for customer base analysis.
  International Journal of Research in Marketing 24(3) 201-209.

Hoppe, Daniel, and Udo Wagner. 'Customer base analysis: The case
  for a central variant of the Betageometric/NBD Model.' Marketing Journal of
  Research and Management 3.2 (2007): 75-90.
}
\seealso{
\code{\link{bgcnbd.Expectation}}
}

