% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.R
\name{mbgnbd.ExpectedCumulativeTransactions}
\alias{mbgnbd.ExpectedCumulativeTransactions}
\title{MBG/NBD Expected Cumulative Transactions}
\usage{
mbgnbd.ExpectedCumulativeTransactions(params, T.cal, T.tot, n.periods.final)
}
\arguments{
\item{params}{MBG/NBD parameters - a vector with \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{T.cal}{a vector to represent customers' calibration period lengths (in
other words, the \code{T.cal} column from a customer-by-sufficient-statistic
matrix).}

\item{T.tot}{end of holdout period. Must be a single value, not a vector.}

\item{n.periods.final}{number of time periods in the calibration and holdout periods.}
}
\value{
Vector of expected cumulative total repeat transactions by all customers.
}
\description{
Calculates the expected cumulative total repeat transactions by all customers
for the calibration and holdout periods.
}
\references{
Batislam, E.P., M. Denizel, A. Filiztekin. 2007. Empirical
  validation and comparison of models for customer base analysis.
  International Journal of Research in Marketing 24(3) 201-209.

Hoppe, Daniel, and Udo Wagner. 'Customer base analysis: The case
  for a central variant of the Betageometric/NBD Model.' Marketing Journal of
  Research and Management 3.2 (2007): 75-90.
}
\seealso{
\code{\link{bgcnbd.ExpectedCumulativeTransactions}}
}

