% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.r
\name{mbgnbd.LL}
\alias{mbgnbd.LL}
\title{Calculates the log-likelihood of the MBG/NBD model}
\usage{
mbgnbd.LL(params, x, t.x, T.cal)
}
\description{
Calculates the log-likelihood of the MBG/NBD model
}
\seealso{
\code{\link{mbgcnbd.LL}}
}

