% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.r
\name{mbgnbd.pmf}
\alias{mbgnbd.pmf}
\title{MBG/NBD Unconditional Probability Distribution of Transactions}
\usage{
mbgnbd.pmf(params, t, x)
}
\arguments{
\item{params}{MBG/NBD parameters - a vector with \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{t}{length of time for which we are calculating the expected number of 
transactions.}

\item{x}{number of transactions for which probability is calculated.}
}
\value{
P(X(t)=x|r,alpha,a,b). If any of the input parameters has a length
  greater than 1, this will be a vector of expected number of transactions.
}
\description{
Uses MBG/NBD model parameters to return the probability distribution of
purchase frequencies for a random customer in a given time period, i.e.
P(X(t)=x|r,alpha,a,b)
}
\references{
Batislam, E.P., M. Denizel, A. Filiztekin. 2007. Empirical
  validation and comparison of models for customer base analysis.
  International Journal of Research in Marketing 24(3) 201-209. - Hoppe,
  Daniel, and Udo Wagner. "Customer base analysis: The case for a central
  variant of the Betageometric/NBD Model." Marketing Journal of Research and
  Management 3.2 (2007): 75-90.
}
\seealso{
\code{\link{bgcnbd.pmf}}

\code{\link{mbgnbd.EstimateParameters}}
}

