% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.DrawFutureTransactions}
\alias{mcmc.DrawFutureTransactions}
\title{Samples number of future transactions based on drawn parameters}
\usage{
mcmc.DrawFutureTransactions(cal.cbs, draws, T.star = cal.cbs$T.star,
  size = NULL)
}
\arguments{
\item{cal.cbs}{data.frame with column \code{t.x} and \code{T.cal}}

\item{draws}{MCMC draws returned by \code{\link{pnbd.mcmc.DrawParameters}}, \code{\link{pggg.mcmc.DrawParameters}}, \code{\link{abe.mcmc.DrawParameters}}}

\item{T.star}{length of period for which future transactions are counted}

\item{size}{number of samples to draw; defaults to the same number of parameter draws that are passed to `draws`}
}
\value{
2-dim array [draw x cust] with sampled future transactions
}
\description{
Samples number of future transactions based on drawn parameters
}
\examples{
cbs <- cdnow.sample()$cbs 
param.draws <- pnbd.mcmc.DrawParameters(cbs, mcmc = 200, burnin = 100, thin = 20, chains = 1)
xstar.draws <- mcmc.DrawFutureTransactions(cbs, param.draws)
}

