% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.ExpectedCumulativeTransactions}
\alias{mcmc.ExpectedCumulativeTransactions}
\title{Expected Cumulative Transactions for Pareto/GGG, Pareto/NBD (HB) and Pareto/NBD (Abe)}
\usage{
mcmc.ExpectedCumulativeTransactions(draws, T.cal, T.tot, n.periods.final)
}
\arguments{
\item{draws}{MCMC draws returned by \code{\link{pnbd.mcmc.DrawParameters}},
\code{\link{pggg.mcmc.DrawParameters}} or \code{\link{abe.mcmc.DrawParameters}}}

\item{T.cal}{a vector to represent customers' calibration period lengths (in
other words, the \code{T.cal} column from a customer-by-sufficient-statistic
matrix). Considering rounding in order to speed up calculations.}

\item{T.tot}{end of holdout period. Must be a single value, not a vector.}

\item{n.periods.final}{number of time periods in the calibration and holdout periods.}
}
\value{
Vector of expected cumulative total repeat transactions by all customers.
}
\description{
Calculates the expected cumulative total repeat transactions by all customers
for the calibration and holdout periods.
}
\examples{
cbs <- cdnow.sample()$cbs
param.draws <- pnbd.mcmc.DrawParameters(cbs, mcmc = 200, burnin = 100, thin = 20, chains = 1) # we use short MCMC runs here only for demo purposes
# Returns a vector containing cumulative repeat transactions for 546 days.
# All parameters are in weeks; the calibration period lasted 39 weeks
# and the holdout period another 39.
mcmc.ExpectedCumulativeTransactions(draws, T.cal = cbs$T.cal, T.tot = 78, n.periods.final = 78)
}
\seealso{
\code{\link{bgcnbd.ExpectedCumulativeTransactions}}
}

