% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.PActive}
\alias{mcmc.PActive}
\title{Calculates P(active) based on drawn future transactions.}
\usage{
mcmc.PActive(xstar)
}
\arguments{
\item{xstar}{future transaction draws returned by \code{\link{mcmc.DrawFutureTransactions}}}
}
\value{
vector with P(active) estimates for each customer
}
\description{
Calculates P(active) based on drawn future transactions.
}
\examples{
cbs <- cdnow.sample()$cbs
param.draws <- pnbd.mcmc.DrawParameters(cbs, mcmc = 3000, burnin = 0, chains = 4)
xstar.draws <- mcmc.DrawFutureTransactions(cbs, param.draws)
mcmc.PActive(xstar.draws)
}

