% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.PAlive}
\alias{mcmc.PAlive}
\title{Calculates P(alive) based on MCMC draws}
\usage{
mcmc.PAlive(cal.cbs, draws)
}
\arguments{
\item{cal.cbs}{data.frame with column \code{T.cal}}

\item{draws}{MCMC draws returned by \code{\link{pnbd.mcmc.DrawParameters}}, \code{\link{pggg.mcmc.DrawParameters}}, \code{\link{abe.mcmc.DrawParameters}}}
}
\value{
numeric vector of probabilities
}
\description{
Calculates P(alive) based on MCMC draws
}
\examples{
cbs <- cdnow.sample()$cbs
param.draws <- pnbd.mcmc.DrawParameters(cbs, mcmc = 3000, burnin = 0, chains = 4)
mcmc.PAlive(cbs, param.draws)
}

