% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.PlotFrequencyInCalibration}
\alias{mcmc.PlotFrequencyInCalibration}
\title{Frequency in Calibration Period for Pareto/GGG, Pareto/NBD (HB) and Pareto/NBD (Abe)}
\usage{
mcmc.PlotFrequencyInCalibration(draws, cal.cbs, censor = 7,
  xlab = "Calibration period transactions", ylab = "Customers",
  title = "Frequency of Repeat Transactions")
}
\arguments{
\item{draws}{MCMC draws returned by \code{\link{pnbd.mcmc.DrawParameters}},
\code{\link{pggg.mcmc.DrawParameters}} or \code{\link{abe.mcmc.DrawParameters}}}

\item{cal.cbs}{calibration period CBS (customer by sufficient statistic). It
must contain columns for frequency ('x') and total time observed ('T.cal').}

\item{censor}{integer used to censor the data.}

\item{xlab}{descriptive label for the x axis.}

\item{ylab}{descriptive label for the y axis.}

\item{title}{title placed on the top-center of the plot.}
}
\value{
Calibration period repeat transaction frequency comparison matrix (actual vs. expected).
}
\description{
Plots a histogram and returns a matrix comparing the actual and expected
number of customers who made a certain number of repeat transactions in the
calibration period, binned according to calibration period frequencies.
}
\details{
The method \code{\link{mcmc.pmf}} is called to calculate the expected numbers based on the corresponding model.
}
\examples{
cbs <- cdnow.sample()$cbs
param.draws <- pnbd.mcmc.DrawParameters(cbs, mcmc = 200, burnin = 100, thin = 20, chains = 1) # we use short MCMC runs here only for demo purposes
mcmc.PlotFrequencyInCalibration(param.draws, cbs)
}
\seealso{
\code{\link[BTYD]{bgnbd.PlotFrequencyInCalibration}} \code{\link{mcmc.pmf}}
}

