% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nbd.cbs.LL}
\alias{nbd.cbs.LL}
\title{Calculate the log-likelihood of the NBD model}
\usage{
nbd.cbs.LL(params, cal.cbs)
}
\arguments{
\item{params}{NBD parameters - a vector with r and alpha, in that
order.}

\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
'x' and total time observed 'T.cal'. Optionally a column 'custs' can be
provided, which represents number of customers with a specific combination
of frequency 'x' and 'T.cal' data.frame with columns 'x' and 'T.cal' and
optional 'custs'.}
}
\value{
the total log-likelihood for the provided data.
}
\description{
Calculate the log-likelihood of the NBD model
}
\seealso{
nbd.EstimateParameters
}

