% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nbd.ConditionalExpectedTransactions}
\alias{nbd.ConditionalExpectedTransactions}
\title{NBD Conditional Expected Transactions}
\usage{
nbd.ConditionalExpectedTransactions(params, T.star, x, T.cal)
}
\arguments{
\item{params}{NBD parameters - a vector with r and alpha, in that order.}

\item{T.star}{length of time for which we are calculating the expected number
of transactions.}

\item{x}{number of repeat transactions in the calibration period T.cal, or a
vector of calibration period frequencies.}

\item{T.cal}{length of calibration period, or a vector of calibration period
lengths.}
}
\value{
Number of transactions a customer is expected to make in a time
  period of length t, conditional on their past behavior. If any of the input
  parameters has a length greater than 1, this will be a vector of expected
  number of transactions.
}
\description{
Uses NBD model parameters and a customer's past transaction behavior to
return the number of transactions they are expected to make in a given time
period.
}
\seealso{
nbd.EstimateParameters
}

