% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbd.R
\name{nbd.GenerateData}
\alias{nbd.GenerateData}
\title{Simulate data according to NBD model assumptions}
\usage{
nbd.GenerateData(n, T.cal, T.star, params, return.elog = FALSE)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{NBD parameters - a vector with \code{r} and \code{alpha} in that order}

\item{return.elog}{boolean - if \code{TRUE} then the event log is returned in
addition to the CBS summary}
}
\value{
list with elements \code{cbs} and \code{elog} containing data.frames
}
\description{
Simulate data according to NBD model assumptions
}
\examples{
n <- 1000  # no. of customers
T.cal <- 32  # length of calibration period
T.star <- 32  # length of hold-out period
params <- c(r = 0.85, alpha = 4.45)  # purchase frequency lambda_i ~ Gamma(r, alpha)
data <- nbd.GenerateData(n, T.cal, T.star, params)
cbs <- data$cbs  # customer by sufficient summary statistic - one row per customer
elog <- data$elog  # Event log - one row per event/purchase
}

