% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbd-mcmc.R
\name{nbd.mcmc.DrawParameters}
\alias{nbd.mcmc.DrawParameters}
\title{NBD (HB) Parameter Draws}
\usage{
nbd.mcmc.DrawParameters(cal.cbs, mcmc = 2500, burnin = 500, thin = 50,
  chains = 2, mc.cores = NULL, param_init = NULL, trace = 100)
}
\arguments{
\item{cal.cbs}{Calibration period customer-by-sufficient-statistic (CBS)
data.frame. It must contain a row for each customer, and columns \code{x}
for frequency,  and \code{T.cal} for the total time observed. A correct
format can be easily generated based on the complete event log of a
customer cohort with \code{\link{elog2cbs}}.}

\item{mcmc}{Number of MCMC steps.}

\item{burnin}{Number of initial MCMC steps which are discarded.}

\item{thin}{Only every \code{thin}-th MCMC step will be returned.}

\item{chains}{Number of MCMC chains to be run.}

\item{mc.cores}{Number of cores to use in parallel (Unix only). Defaults to \code{min(chains, detectCores())}.}

\item{param_init}{List of start values for cohort-level parameters.}

\item{trace}{Print logging statement every \code{trace}-th iteration. Not available for \code{mc.cores > 1}.}
}
\value{
2-element list:
\itemize{
 \item{\code{level_1 }}{list of \code{\link{mcmc.list}}s, one for each customer, with draws for customer-level parameters \code{lambda}}
 \item{\code{level_2 }}{\code{\link{mcmc.list}}, with draws for cohort-level parameters \code{r}, \code{alpha}}
}
}
\description{
Returns draws from the posterior distributions of the NBD (HB)
parameters, on cohort as well as on customer level.
}
\examples{
data("groceryElog")
cbs <- elog2cbs(groceryElog, T.cal = "2006-12-31")
param.draws <- mcmc.DrawParameters(cbs,
  mcmc = 200, burnin = 100, thin = 20, chains = 1) # short MCMC to run demo fast

# cohort-level parameter draws
as.matrix(param.draws$level_2)
# customer-level parameter draws for customer with ID '4'
as.matrix(param.draws$level_1[["4"]])
}
\seealso{
\code{\link{nbd.GenerateData} }
}
