% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pcnbd.mcmc.DrawFutureTransactions}
\alias{pcnbd.mcmc.DrawFutureTransactions}
\title{Samples number of future transactions based on drawn parameters}
\usage{
pcnbd.mcmc.DrawFutureTransactions(cal.cbs, draws, T.star = cal.cbs$T.star)
}
\arguments{
\item{cal.cbs}{data.frame with column \code{t.x} and \code{T.cal}}

\item{draws}{MCMC draws returned by \code{\link{pcnbd.mcmc.DrawParameters}}}

\item{T.star}{length of period for which future transactions are counted}
}
\value{
2-dim array [draw x cust] with sampled future transactions
}
\description{
Samples number of future transactions based on drawn parameters
}

