% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pcnbd.mcmc.DrawParameters}
\alias{pcnbd.mcmc.DrawParameters}
\title{Hierarchical Bayes implementation of Pareto/CNBD}
\usage{
pcnbd.mcmc.DrawParameters(cal.cbs, mcmc = 1500, burnin = 500, thin = 1,
  chains = 2, param_init = list(t = 1, gamma = 1, r = 1, alpha = 1, s = 1,
  beta = 1), hyper_prior = list(t_1 = 1/1000, t_2 = 1/1000, gamma_1 = 1/1000,
  gamma_2 = 1/1000, r_1 = 1/1000, r_2 = 1/1000, alpha_1 = 1/1000, alpha_2 =
  1/1000, s_1 = 1/1000, s_2 = 1/1000, beta_1 = 1/1000, beta_2 = 1/1000))
}
\arguments{
\item{cal.cbs}{data.frame with columns 'x', 't.x', 'T.cal', 'litt'; e.g. output of \code{\link{elog2cbs}}}

\item{mcmc}{number of MCMC steps}

\item{burnin}{number of initial MCMC steps which are discarded}

\item{thin}{only every thin-th MCMC step will be returned}

\item{chains}{number of MCMC chains to be run}

\item{use_data_augmentation}{determines MCMC method to be used}

\item{param_init}{list of 2nd-level parameter start values}

\item{hyper_prior}{list of hyper parameters for 2nd-level parameters}
}
\value{
2-element list
level_1:  list of coda::mcmc.list objects; one for each customer, containing individual-level draws
level_2:  coda::mcmc.list object containing draws of heterogeneity parameters
}
\description{
Returns 2-element list
  level_1:  3-dim array [draw x parameter x cust] wrapped as coda::mcmc.list object
  level_2:  2-dim array [draw x parameter] wrapped as coda::mcmc.list object
}
\examples{
#params <- list(r=1.4, alpha=1.3, s=0.7, beta=7, t=1.5, gamma=1)
#cbs <- pcnbd.GenerateData(1000, 10, 5, params)$cbs
#draws <- pcnbd.mcmc.DrawParameters(cbs, mcmc=1000, chains=2)
#plot(draws$level_2, ask=F)
#rbind("actual"=unlist(params), "estimated"=summary(draws$level_2, quantiles=0.5)$quantiles)
}
\seealso{
pcnbd.GenerateData elog2cbs
}

