% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pcnbd.mcmc.PAlive}
\alias{pcnbd.mcmc.PAlive}
\title{Calculates P(alive) based on MCMC draws}
\usage{
pcnbd.mcmc.PAlive(data, draws)
}
\arguments{
\item{data}{data.frame with column 'T.cal'}

\item{draws}{MCMC draws returned by \code{pcnbd.mcmc.DrawParameters}}
}
\value{
numeric vector of probabilities
}
\description{
Calculates P(alive) based on MCMC draws
}

