% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pcnbd.mcmc.plotRegularityRateHeterogeneity}
\alias{pcnbd.mcmc.plotRegularityRateHeterogeneity}
\title{Pareto/CNBD Plot Regularity Rate Heterogeneity}
\usage{
pcnbd.mcmc.plotRegularityRateHeterogeneity(draws, xmax = NULL)
}
\arguments{
\item{draws}{MCMC draws returned by \code{\link{pcnbd.mcmc.DrawParameters}}}

\item{xmax}{upper bound for x-scale}
}
\description{
Plots and returns the estimated gamma distribution of k (customers' regularity in interpurchase times).
}

