% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto-ggg-mcmc.R
\name{pggg.mcmc.plotRegularityRateHeterogeneity}
\alias{pggg.mcmc.plotRegularityRateHeterogeneity}
\title{Pareto/GGG Plot Regularity Rate Heterogeneity}
\usage{
pggg.mcmc.plotRegularityRateHeterogeneity(draws, xmax = NULL, fn = NULL)
}
\arguments{
\item{draws}{MCMC draws returned by \code{\link{pggg.mcmc.DrawParameters}}}

\item{xmax}{upper bound for x-scale}

\item{fn}{optional function to summarize individual-level draws for k, e.g. 'mean'}
}
\description{
Plots and returns the estimated gamma distribution of k (customers' regularity in interpurchase times).
}

