% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plotSampledTimingPatterns}
\alias{plotSampledTimingPatterns}
\title{Plot timing patterns of sampled customers}
\usage{
plotSampledTimingPatterns(elog, T.cal = NULL, n = 30,
  title = "Sampled Timing Patterns")
}
\arguments{
\item{elog}{Event log, a \code{data.frame} with columns \code{cust} and
transaction time \code{t} or \code{date}.}

\item{T.cal}{End of calibration period, which is visualized as a vertical line.}

\item{n}{Number of sampled customers.}

\item{title}{Plot title.}
}
\description{
Plot timing patterns of sampled customers
}
\examples{
data("groceryElog")
plotSampledTimingPatterns(groceryElog, T.cal = "2006-12-31")
}

