% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto-nbd-mcmc.R
\name{pnbd.GenerateData}
\alias{pnbd.GenerateData}
\title{Generate artificial data which follows Pareto/NBD model assumptions}
\usage{
pnbd.GenerateData(n, T.cal, T.star, params, return.elog = FALSE)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{list of parameters: {r, alpha, s, beta}}

\item{return.elog}{if \code{TRUE} then the event-log is returned as well; decreases performance}
}
\value{
2-elemnt list
}
\description{
Returns 2-element list
* cbs: data.frame with 'cust', \code{x}, \code{t.x}, \code{T.cal}, 'T.star', 'x.star' 
       this is the summary statistics data.frame which contains all 
       needed information for parameter estimation
* elog: data.frame with 'cust', \code{t}
}
\examples{
params <- list(r = 5, alpha = 10, s = 0.8, beta = 12)
data <- pnbd.GenerateData(n = 1000, T.cal = 32, T.star = 32, params, return.elog = TRUE)
cbs <- data$cbs  # customer by sufficient summary statistic - one row per customer
elog <- data$elog  # Event log - one row per event/purchase
}

