% Generated by roxygen2 (4.0.1): do not edit by hand
\name{abe.mcmc.DrawParameters}
\alias{abe.mcmc.DrawParameters}
\title{HB Pareto/NBD variant as described in Abe (2009)}
\usage{
abe.mcmc.DrawParameters(cal.cbs, covariates = c(), mcmc = 1500,
  burnin = 500, thin = 50, chains = 2, trace = 100)
}
\arguments{
\item{cal.cbs}{data.frame with columns \code{x}, \code{t.x}, \code{T.cal}}

\item{covariates}{list of column names containing customer-level covariates}

\item{mcmc}{number of MCMC steps}

\item{burnin}{number of initial MCMC steps which are discarded}

\item{thin}{only every thin-th MCMC step will be returned}

\item{chains}{number of MCMC chains to be run}

\item{trace}{print logging step every \code{trace} iteration}
}
\value{
2-element list:
\itemize{
 \item{\code{level_1}}{list of \code{\link{mcmc.list}} objects; one for each customer, containing individual-level draws}
 \item{\code{level_2}}{\code{\link{mcmc.list}} object containing draws of heterogeneity parameters}
}
}
\description{
Returns 2-element list
  level_1:  3-dim array [draw x parameter x cust] wrapped as coda::mcmc.list object
  level_2:  2-dim array [draw x parameter] wrapped as coda::mcmc.list object
}
\examples{

# Load CDNow event log
data(cdnowElog, package="BTYD")
elog <- data.frame(t(sapply(2:nrow(cdnowElog), function(i) strsplit(as.character(cdnowElog[i,]), split=",")[[1]])), stringsAsFactors=FALSE)
names(elog) <- c("cust", "sampleid", "date", "cds", "sales")
elog$date <- as.Date(elog$date, "\%Y\%m\%d")
elog$sales <- as.numeric(elog$sales)

# Transform to CBS (including extra column 'litt')
cbs <- elog2cbs(elog, per="week", T.cal=as.Date("1997-09-30"))

# Estimate with no covariates, i.e. model M1 in (Abe 2009)
set.seed(1)
draws_m1 <- abe.mcmc.DrawParameters(cbs, mcmc=5000, burnin=5000)
plot(draws_m1$level_2, density=FALSE)
params.pnbd_abe.m1 <- round(summary(draws_m1$level_2)$quantiles[, c("2.5\%", "50\%", "97.5\%")], 2)
params.pnbd_abe.m1
# -> Parameter Estimates match (Abe 2009)

# Append dollar amount of first purchase; used as covariate in (Abe 2009)
elog.dt <- data.table(elog)
setkey(elog.dt, cust, date)
cbs <- merge(cbs, elog.dt[, list(first=as.numeric(sales[1])*10^-3), by="cust"], by="cust")

set.seed(1)
draws_m2 <- abe.mcmc.DrawParameters(cbs, covariates=c("first"), mcmc=5000, burnin=5000)
plot(draws_m2$level_2, density=FALSE)
params.pnbd_abe.m2 <- round(summary(draws_m2$level_2)$quantiles[, c("2.5\%", "50\%", "97.5\%")], 4)
params.pnbd_abe.m2
# -> Parameter Estimates match (Abe 2009)
}
\references{
Abe, Makoto. "Counting your customers one by one: A hierarchical Bayes extension to the Pareto/NBD model." Marketing Science 28.3 (2009): 541-553.
}
\seealso{
\code{link{abe.GenerateData}} \code{\link{pcnbd.mcmc.PAlive}} \code{\link{pcnbd.mcmc.DrawFutureTransactions}}
}

