% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto-nbd-abe.R
\name{abe.mcmc.DrawParameters}
\alias{abe.mcmc.DrawParameters}
\title{HB Pareto/NBD variant as described in Abe (2009)}
\usage{
abe.mcmc.DrawParameters(cal.cbs, covariates = c(), mcmc = 1500,
  burnin = 500, thin = 50, chains = 2, mc.cores = NULL, trace = 100)
}
\arguments{
\item{cal.cbs}{data.frame with columns \code{x}, \code{t.x}, \code{T.cal}}

\item{covariates}{list of column names containing customer-level covariates}

\item{mcmc}{number of MCMC steps}

\item{burnin}{number of initial MCMC steps which are discarded}

\item{thin}{only every thin-th MCMC step will be returned}

\item{chains}{number of MCMC chains to be run}

\item{mc.cores}{number of cores to use in parallel (Unix only); defaults to min(chains, detectCores())}

\item{trace}{print logging step every \code{trace} iteration}
}
\value{
2-element list:
\itemize{
 \item{\code{level_1}}{list of \code{\link{mcmc.list}} objects; one for each customer, containing individual-level draws}
 \item{\code{level_2}}{\code{\link{mcmc.list}} object containing draws of heterogeneity parameters}
}
}
\description{
Returns 2-element list
  level_1:  3-dim array [draw x parameter x cust] wrapped as coda::mcmc.list object
  level_2:  2-dim array [draw x parameter] wrapped as coda::mcmc.list object
}
\details{
See \code{demo('pareto-nbd-abe')} for how to use this model.
}
\references{
Abe, Makoto. 'Counting your customers one by one: A hierarchical Bayes extension to the Pareto/NBD model.' Marketing Science 28.3 (2009): 541-553.
}
\seealso{
\code{link{abe.GenerateData}} \code{\link{mcmc.PAlive}} \code{\link{mcmc.DrawFutureTransactions}}
}

