% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.r
\name{bgcnbd.GenerateData}
\alias{bgcnbd.GenerateData}
\title{Simulate data according to BG/CNBD-k model assumptions}
\usage{
bgcnbd.GenerateData(n, T.cal, T.star = T.cal, params, return.elog = FALSE,
  dropout_at_zero = FALSE)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{return.elog}{boolean - if \code{TRUE} then the event log is returned in 
addition to the CBS summary}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
list with elements \code{cbs} and \code{elog} containing data.frames
}
\description{
Simulate data according to BG/CNBD-k model assumptions
}

