% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-cnbd-k.R
\name{bgcnbd.PAlive}
\alias{bgcnbd.PAlive}
\title{BG/CNBD-k P(alive)}
\usage{
bgcnbd.PAlive(params, x, t.x, T.cal, dropout_at_zero = FALSE)
}
\arguments{
\item{params}{BG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a}
and \code{b} in that order.}

\item{x}{number of repeat transactions in the calibration period T.cal, or a 
vector of calibration period frequencies.}

\item{t.x}{recency, i.e. length between first and last transaction during 
calibration period.}

\item{T.cal}{length of calibration period, or a vector of calibration period 
lengths.}

\item{dropout_at_zero}{Boolean; the mbg-methods are simple wrapper methods,
which set this parameter to TRUE}
}
\value{
Probability that the customer is still alive at the end of the 
  calibration period.
}
\description{
Uses BG/CNBD-k model parameters and a customer's past transaction behavior 
to return the probability that they are still alive at the end of the 
calibration period.
}
\examples{
cbs <- cdnow.sample()$cbs # load CDNow summary data
params <- bgcnbd.EstimateParameters(cbs)
palive <- bgcnbd.PAlive(params, cbs$x, cbs$t.x, cbs$T.cal)
mean(palive) # Estimated share of customers to be still alive
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

