% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-nbd.R
\name{bgnbd.GenerateData}
\alias{bgnbd.GenerateData}
\title{Simulate data according to BG/NBD model assumptions}
\usage{
bgnbd.GenerateData(n, T.cal, T.star, params, return.elog = FALSE)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{BG/NBD parameters - a vector with \code{r}, \code{alpha}, \code{a} and
\code{b} in that order.}

\item{return.elog}{boolean - if \code{TRUE} then the event log is returned in
addition to the CBS summary}
}
\value{
list with elements \code{cbs} and \code{elog} containing data.frames
}
\description{
Simulate data according to BG/NBD model assumptions
}

