% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bgnbd.LL}
\alias{bgnbd.LL}
\title{Calculate the log-likelihood of the BG/NBD model}
\usage{
bgnbd.LL(params, x, t.x, T.cal)
}
\arguments{
\item{params}{BG/NBD parameters - a vector with \code{r}, \code{alpha}, \code{a} and
\code{b} in that order.}

\item{x}{frequency, i.e. number of re-purchases}

\item{t.x}{recency, i.e. time elapsed from first purchase to last purchase}

\item{T.cal}{total time of observation period}
}
\value{
a vector of log-likelihoods
}
\description{
Calculate the log-likelihood of the BG/NBD model
}
\seealso{
\code{\link{bgnbd.EstimateParameters}}
}

