% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bgnbd.PAlive}
\alias{bgnbd.PAlive}
\title{BG/NBD P(alive)}
\usage{
bgnbd.PAlive(params, x, t.x, T.cal)
}
\arguments{
\item{params}{BG/NBD parameters - a vector with 'r', 'alpha', 'a' and 'b' in
that order.}

\item{x}{number of repeat transactions in the calibration period T.cal, or a
vector of calibration period frequencies.}

\item{t.x}{recency, i.e. length between first and last transaction during
calibration period.}

\item{T.cal}{length of calibration period, or a vector of calibration period
lengths.}
}
\value{
Probability that the customer is still alive at the end of the calibration period.
}
\description{
Uses BG/NBD model parameters and a customer's past transaction behavior to
return the probability that they are still alive at the end of the
calibration period.
}

