% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cbgcnbd.cbs.LL}
\alias{cbgcnbd.cbs.LL}
\title{Calculate the log-likelihood of the CBG/CNBD-k model}
\usage{
cbgcnbd.cbs.LL(params, cal.cbs)
}
\arguments{
\item{params}{CBG/CNBD-k parameters - a vector with 'k', \code{r}, \code{alpha}, \code{a}
and \code{b} in that order.}

\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
\code{x}, for recency \code{t.x}, for sum of logarithmic interpurchase times 'litt'
and total time observed \code{T.cal}. Optionally a column \code{custs} can be
provided, which represents number of customers with a specific combination
of frequency \code{x} and \code{T.cal}.}
}
\value{
the total log-likelihood for the provided data.
}
\description{
Calculate the log-likelihood of the CBG/CNBD-k model
}
\seealso{
\code{\link{cbgcnbd.EstimateParameters}}
}

