% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cbgcnbd.GenerateData}
\alias{cbgcnbd.GenerateData}
\title{Simulate data according to CBG/CNBD-k model assumptions}
\usage{
cbgcnbd.GenerateData(n, T.cal, T.star = T.cal, params, return.elog = FALSE)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{CBG/CNBD-k parameters - a vector with 'k', \code{r}, \code{alpha}, \code{a} and \code{b}
in that order.}

\item{return.elog}{boolean - if \code{TRUE} then the event log is returned in
addition to the CBS summary}
}
\value{
list with elements \code{cbs} and \code{elog} containing data.frames
}
\description{
Simulate data according to CBG/CNBD-k model assumptions
}

