% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cbgcnbd.LL}
\alias{cbgcnbd.LL}
\title{Calculate the log-likelihood of the CBG/CNBD-k model}
\usage{
cbgcnbd.LL(params, x, t.x, T.cal, litt)
}
\arguments{
\item{params}{CBG/CNBD-k parameters - a vector with 'k', \code{r}, \code{alpha}, \code{a}
and \code{b} in that order.}

\item{x}{frequency, i.e. number of re-purchases}

\item{t.x}{recency, i.e. time elapsed from first purchase to last purchase}

\item{T.cal}{total time of observation period}

\item{litt}{sum of logarithmic interpurchase times}
}
\value{
a vector of log-likelihoods
}
\description{
Calculate the log-likelihood of the CBG/CNBD-k model
}
\seealso{
\code{\link{cbgcnbd.EstimateParameters}}
}

