% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cbgnbd.GenerateData}
\alias{cbgnbd.GenerateData}
\title{Simulate data according to CBG/NBD model assumptions}
\usage{
cbgnbd.GenerateData(n, T.cal, T.star, params, return.elog = FALSE)
}
\description{
Simulate data according to CBG/NBD model assumptions
}
\seealso{
\code{\link{cbgcnbd.GenerateData}}
}

