% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cdnow.sample}
\alias{cdnow.sample}
\title{CDNow Sample Data}
\usage{
cdnow.sample(T.cal = "1997-09-30", T.tot = "1998-06-30")
}
\arguments{
\item{T.cal}{Final date of calibration period.}

\item{T.tot}{Final date of holdout period.}
}
\value{
List of length 2:
\item{\code{cbs}}{A data.frame with a row for each customer and the summary statistic as columns, generated with \code{\link{elog2cbs}}}
\item{\code{elog}}{A data.frame with a row for each transaction, and columns \code{cust}, \code{date}, \code{sales} and \code{cds}.}
}
\description{
This is a convenience wrapper for \code{data('cdnowElog', package='BTYD')},
with same-day transactions being merged together for each customer.
}
\examples{
cdnow <- cdnow.sample()
head(cdnow$cbs)
head(cdnow$elog)
}
\references{
Fader, Peter S. and Bruce G.,S. Hardie, (2001), 'Forecasting
  Repeat Sales at CDNOW: A Case Study,' Interfaces, 31 (May-June), Part 2 of
  2, S94-S107.
}

