% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dc.PlotTracking}
\alias{dc.PlotTracking}
\title{Generic Method for Tracking Plots}
\usage{
dc.PlotTracking(
  actual,
  expected,
  T.cal = NULL,
  xlab = "",
  ylab = "",
  title = "",
  xticklab = NULL,
  ymax = NULL,
  legend = c("Actual", "Model")
)
}
\description{
Generic Method for Tracking Plots
}
\keyword{internal}
