% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{elog2cbs}
\alias{elog2cbs}
\title{Faster implementation of BTYD::dc.ElogToCbsCbt that also returns summary statistic for estimating regularity}
\usage{
elog2cbs(elog, per = "week", T.cal = max(elog$date),
  T.tot = max(elog$date))
}
\arguments{
\item{elog}{data.frame with columns 'cust' and 'date'; optionally with column 'sales'}

\item{per}{time unit, either 'week', 'day', 'hour', 'min', 'sec'}

\item{T.cal}{end date of calibration period}

\item{T.tot}{end date of holdout period}
}
\value{
data.frame
}
\description{
Returns data.frame with
     cust:   customer id (unique key)
     x:      nr of recurring events in calibration period
     t.x:    time between first and last event in calibration period
     litt:   sum of logarithmic intertransaction timings durint calibration period 
             this is a summary statistic for estimating regularity
     sales:  sum of sales in calibration period
     first:  date of first transaction in calibration period
     T.cal:  time between first event and end of calibration period
     T.star: length of holdout period
     x.star: nr of events within holdout period
     sales.star: sum of sales within holdout period
     
Customers without any transaction during calibration period are being dropped from the result.
Transactions with identical `cust` and `date` field are treated as a single transaction, with `sales` being summed up
}

