% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{elog2cum}
\alias{elog2cum}
\title{Aggregate Event Log to cumulative number of repeat transactions}
\usage{
elog2cum(elog, by = 7)
}
\arguments{
\item{elog}{Event log, a \code{data.frame} with columns \code{cust} and 
transaction time \code{t} or \code{date}.}

\item{by}{Only return every \code{}-th number. Defaults to 7, and thus
returns weekly numbers.}
}
\value{
Numeric vector of cumulative repeat transactions.
}
\description{
Aggregate Event Log to cumulative number of repeat transactions
}
\examples{
elog <- cdnow.sample()$elog
cum <- elog2cum(elog)
plot(cum, typ="l")
}
\seealso{
\code{\link{elog2inc}}
}

