% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{elog2inc}
\alias{elog2inc}
\title{Convert Event Log to (weekly) incremental number of repeat transactions}
\usage{
elog2inc(elog, by = 7)
}
\arguments{
\item{elog}{Event Log}

\item{by}{defaults to 7, which means weekly numbers}
}
\value{
numeric vector
}
\description{
Convert Event Log to (weekly) incremental number of repeat transactions
}
\examples{
elog <- cdnow.sample()$elog
inc <- elog2inc(elog)
plot(inc, typ='l')
}
\seealso{
\code{\link{elog2cum}}
}

