% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{estimateRegularity}
\alias{estimateRegularity}
\title{Estimate Regularity in Intertransaction Timings}
\usage{
estimateRegularity(elog, method = "wheat", plot = FALSE)
}
\arguments{
\item{elog}{data.frame with transaction logs; requires columns customer-id
'cust' and transaction time \code{t} or 'date'}

\item{method}{\code{wheat}, \code{mle}, \code{mle-minka}, \code{mle-thom}, \code{cv}}

\item{plot}{if \code{TRUE} then distribution of estimated regularity will be plotted}
}
\value{
estimated real-valued regularity parameter; rounded to an integer,
  this can be used as \code{k} for estimating MBG/CNBD-k models
}
\description{
Estimates degree of regularity of intertransaction timings of a customer cohort,
}
\details{
This is done
 - by either assuming same regularity across all customers; this then only 
     require three transactions per customer
     method: wheat
 - or by estimating regularity for each customer seperately (as the shape
     parameter of a fitted gamma distribution), and then return the 
     median across estimates; this requires min. 10 transactions per customer
     methods: mle, mle-minka, mle-thom, cv
}
\examples{
# generate Erlang-3 with various lambdas
n <- 100
k <- 3
x <- 20
elog <- do.call(rbind, lapply(1:n, function(i) {
  lambda <- exp(rnorm(1))
    data.frame(cust = i, t = cumsum(rgamma(x, k, k * lambda)))
    }))
    
# estimate regularity parameter k
estimateRegularity(elog, plot = TRUE, method = 'wheat')
estimateRegularity(elog, plot = TRUE, method = 'mle-minka')
estimateRegularity(elog, plot = TRUE, method = 'mle-thom')
estimateRegularity(elog, plot = TRUE, method = 'cv') 
}
\references{
Wheat, Rita D., and Donald G. Morrison.  'Estimating purchase regularity with two interpurchase times.'
}

