% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing.r
\name{estimateRegularity}
\alias{estimateRegularity}
\title{Estimate Regularity in Intertransaction Timings}
\usage{
estimateRegularity(elog, method = "wheat", plot = FALSE)
}
\arguments{
\item{elog}{data.frame with transaction logs; requires columns customer-id
'cust' and transaction time \code{t} or 'date'}

\item{method}{'wheat', 'mle', 'mle-minka', 'mle-thom', 'cv'}

\item{plot}{if \code{TRUE} then distribution of estimated regularity will be plotted}
}
\value{
estimated real-valued regularity parameter; rounded to an integer,
  this can be used as 'k' for estimating MBG/CNBD-k models
}
\description{
Estimates degree of regularity of intertransaction timings of a customer cohort,
}
\details{
This is done
 - by either assuming same regularity across all customers; this then only 
     require three transactions per customer
     method: wheat
 - or by estimating regularity for each customer seperately (as the shape
     parameter of a fitted gamma distribution), and then return the 
     median across estimates; this requires min. 10 transactions per customer
     methods: mle, mle-minka, mle-thom, cv
}
\examples{
require(data.table)

# generate Erlang-3 with various lambdas
nr_of_customers <- 100
elog <- rbindlist(lapply(1:nr_of_customers, function(i) {
  k <- 3
  lambda <- exp(rnorm(1))
  data.table(cust=i, t=cumsum(rgamma(20, k, k*lambda)))
}))

# estimate regularity parameter k
estimateRegularity(elog, plot=T, method='wheat')
estimateRegularity(elog, plot=T, method='mle-minka')
estimateRegularity(elog, plot=T, method='mle-thom')
estimateRegularity(elog, plot=T, method='cv')
}
\seealso{
\code{\link{mbgcnbd.EstimateParameters}}
}

