% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gg-nbd.r
\name{ggnbd.cbs.LL}
\alias{ggnbd.cbs.LL}
\title{Calculate the log-likelihood of the Gamma/Gompertz/NBD model}
\usage{
ggnbd.cbs.LL(params, cal.cbs)
}
\arguments{
\item{params}{Gamma/Gompertz/NBD parameters - a vector with \code{r},
\code{alpha}, \code{b}, \code{s} and \code{beta} in that order.}

\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
\code{x}, for recency \code{t.x} and total time observed \code{T.cal}.
Optionally a column \code{custs} can be provided, which represents number
of customers with a specific combination of frequency \code{x} and
\code{T.cal}.}
}
\value{
the total log-likelihood for the provided data.
}
\description{
Calculate the log-likelihood of the Gamma/Gompertz/NBD model
}
\seealso{
\code{\link{ggnbd.EstimateParameters}}
}

