% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ggnbd.EstimateParameters}
\alias{ggnbd.EstimateParameters}
\title{Parameter Estimation for Gamma/Gompertz/NBD model}
\usage{
ggnbd.EstimateParameters(cal.cbs, par.start = c(1, 1, 0.5, 1, 1),
  min.param.value = 1e-06, max.param.value = 10000, trace = 0)
}
\arguments{
\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
'x', for recency 't.x.' and total time observed 'T.cal'. Optionally a
column 'custs' can be provided, which represents number of customers with a
specific combination of frequency 'x', recency 't.x' and 'T.cal'.}

\item{par.start}{initial Gamma/Gompertz/NBD parameters - a vector with 'r',
'alpha', 'b', 's' and 'beta' in that order.}

\item{max.param.value}{the upper bound on parameters}
}
\value{
list of estimated parameters
}
\description{
Estimates parameters for the Gamma/Gompertz/NBD model via Maximum Likelihood
Estimation.
}
\references{
Bemmaor, Albert C., and Nicolas Glady. "Modeling Purchasing
  Behavior with Sudden “Death”: A Flexible Customer Lifetime Model."
  Management Science 58.5 (2012): 1012-1021.
}

