% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.r
\name{mbgcnbd.EstimateParameters}
\alias{mbgcnbd.EstimateParameters}
\title{Parameter Estimation for the MBG/CNBD-k model}
\usage{
mbgcnbd.EstimateParameters(...)
}
\arguments{
\item{...}{Arguments to be passed to bgcnbd.EstimateParameters}
}
\description{
Parameter Estimation for the MBG/CNBD-k model
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

