% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.ExpectedCumulativeTransactions}
\alias{mbgcnbd.ExpectedCumulativeTransactions}
\title{MBG/CNBD-k Expected Cumulative Transactions}
\usage{
mbgcnbd.ExpectedCumulativeTransactions(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{bgnbd.ExpectedCumulativeTransactions}}}
}
\description{
MBG/CNBD-k Expected Cumulative Transactions
}
\references{
Platzer Michael, and Thomas Reutterer (forthcoming)
}

