% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.R
\name{mbgcnbd.ExpectedCumulativeTransactions}
\alias{bgcnbd.ExpectedCumulativeTransactions}
\alias{mbgcnbd.ExpectedCumulativeTransactions}
\title{(M)BG/CNBD-k Expected Cumulative Transactions}
\usage{
mbgcnbd.ExpectedCumulativeTransactions(params, T.cal, T.tot, n.periods.final)

bgcnbd.ExpectedCumulativeTransactions(params, T.cal, T.tot, n.periods.final)
}
\arguments{
\item{params}{A vector with model parameters \code{k}, \code{r}, 
\code{alpha}, \code{a} and \code{b}, in that order.}

\item{T.cal}{A vector to represent customers' calibration period lengths.}

\item{T.tot}{End of holdout period. Must be a single value, not a vector.}

\item{n.periods.final}{Number of time periods in the calibration and holdout
periods.}
}
\value{
Vector of length \code{n.periods.final} with expected cumulative
  total repeat transactions by all customers.
}
\description{
Calculates the expected cumulative total repeat transactions by all customers
for the calibration and holdout periods.
}
\examples{
cbs <- cdnow.sample()$cbs
params <- mbgcnbd.EstimateParameters(cbs)
# Returns a vector containing cumulative repeat transactions for 546 days.
# All parameters are in weeks; the calibration period lasted 39 weeks
# and the holdout period another 39.
mbgcnbd.ExpectedCumulativeTransactions(params, T.cal = cbs$T.cal, T.tot = 78, n.periods.final = 78)
}

