% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.r
\name{mbgcnbd.PAlive}
\alias{mbgcnbd.PAlive}
\title{MBG/CNBD-k P(alive)}
\usage{
mbgcnbd.PAlive(params, x, t.x, T.cal)
}
\arguments{
\item{params}{MBG/CNBD-k parameters - a vector with \code{k}, \code{r}, \code{alpha}, \code{a}
and \code{b} in that order.}

\item{x}{number of repeat transactions in the calibration period T.cal, or a 
vector of calibration period frequencies.}

\item{t.x}{recency, i.e. length between first and last transaction during 
calibration period.}

\item{T.cal}{length of calibration period, or a vector of calibration period 
lengths.}
}
\value{
Probability that the customer is still alive at the end of the 
  calibration period.
}
\description{
Uses MBG/CNBD-k model parameters and a customer's past transaction behavior 
to return the probability that they are still alive at the end of the 
calibration period.
}
\seealso{
\code{\link{bgcnbd.PAlive}}

\code{\link{mbgcnbd.EstimateParameters}}
}

