% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.r
\name{mbgcnbd.PlotTrackingCum}
\alias{mbgcnbd.PlotTrackingCum}
\title{MBG/CNBD-k Tracking Cumulative Transactions Plot}
\usage{
mbgcnbd.PlotTrackingCum(...)
}
\arguments{
\item{...}{Arguments to be passed to bgcnbd.PlotTrackingCum}
}
\description{
MBG/CNBD-k Tracking Cumulative Transactions Plot
}
\seealso{
\code{\link{bgcnbd.PlotTrackingCum}}
}

