% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.r
\name{mbgcnbd.PlotTrackingInc}
\alias{mbgcnbd.PlotTrackingInc}
\title{MBG/CNBD-k Tracking Incremental Transactions Comparison}
\usage{
mbgcnbd.PlotTrackingInc(...)
}
\arguments{
\item{...}{Arguments to be passed to bgcnbd.PlotTrackingInc}
}
\description{
MBG/CNBD-k Tracking Incremental Transactions Comparison
}
\seealso{
\code{\link{bgcnbd.PlotTrackingInc}}
}

