% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-cnbd-k.r
\name{mbgcnbd.pmf}
\alias{mbgcnbd.pmf}
\title{MBG/CNBD-k Unconditional Probability Distribution of Transactions}
\usage{
mbgcnbd.pmf(...)
}
\arguments{
\item{...}{Arguments to be passed to bgcnbd.pmf}
}
\description{
MBG/CNBD-k Unconditional Probability Distribution of Transactions
}
\seealso{
\code{\link{bgcnbd.pmf}}
}

