% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.R
\name{mbgnbd.cbs.LL}
\alias{mbgnbd.cbs.LL}
\title{Calculate the log-likelihood of the MBG/NBD model}
\usage{
mbgnbd.cbs.LL(params, cal.cbs)
}
\arguments{
\item{params}{MBG/NBD parameters - a vector with \code{r}, \code{alpha}, \code{a} 
and \code{b} in that order.}

\item{cal.cbs}{calibration period CBS. It must contain columns for frequency 
\code{x}, for recency \code{t.x}, for sum of logarithmic interpurchase times \code{litt}
and total time observed \code{T.cal}. Optionally a column \code{custs} can be
provided, which represents number of customers with a specific combination
of frequency \code{x} and \code{T.cal}.}
}
\value{
the total log-likelihood for the provided data.
}
\description{
Calculate the log-likelihood of the MBG/NBD model
}
\references{
Batislam, E.P., M. Denizel, A. Filiztekin. 2007. Empirical
  validation and comparison of models for customer base analysis.
  International Journal of Research in Marketing 24(3) 201-209. - Hoppe,
  Daniel, and Udo Wagner. 'Customer base analysis: The case for a central
  variant of the Betageometric/NBD Model.' Marketing Journal of Research and
  Management 3.2 (2007): 75-90.
}
\seealso{
\code{\link{bgcnbd.cbs.LL}}

\code{\link{mbgnbd.EstimateParameters}}
}

