% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.R
\name{mbgnbd.ConditionalExpectedTransactions}
\alias{mbgnbd.ConditionalExpectedTransactions}
\title{MBG/NBD Conditional Expected Transactions}
\usage{
mbgnbd.ConditionalExpectedTransactions(params, T.star, x, t.x, T.cal)
}
\arguments{
\item{params}{MBG/NBD parameters - a vector with \code{r}, \code{alpha}, \code{a}
and \code{b} in that order.}

\item{T.star}{length of time for which we are calculating the expected number
of transactions.}

\item{x}{number of repeat transactions in the calibration period T.cal, or a 
vector of calibration period frequencies.}

\item{t.x}{recency, i.e. length between first and last transaction during 
calibration period.}

\item{T.cal}{length of calibration period, or a vector of calibration period 
lengths.}
}
\value{
Number of transactions a customer is expected to make in a time 
  period of length t, conditional on their past behavior. If any of the input
  parameters has a length greater than 1, this will be a vector of expected 
  number of transactions.
}
\description{
Uses MBG/NBD model parameters and a customer's past transaction behavior 
to return the number of transactions they are expected to make in a given 
time period.
}
\references{
Batislam, E.P., M. Denizel, A. Filiztekin. 2007. Empirical
  validation and comparison of models for customer base analysis.
  International Journal of Research in Marketing 24(3) 201-209. - Hoppe,
  Daniel, and Udo Wagner. "Customer base analysis: The case for a central
  variant of the Betageometric/NBD Model." Marketing Journal of Research and
  Management 3.2 (2007): 75-90.
}
\seealso{
\code{\link{bgcnbd.ConditionalExpectedTransactions}}

\code{\link{mbgnbd.EstimateParameters}}
}

