% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.R
\name{mbgnbd.EstimateParameters}
\alias{mbgnbd.EstimateParameters}
\title{Parameter Estimation for the MBG/NBD model}
\usage{
mbgnbd.EstimateParameters(cal.cbs, par.start = c(1, 1, 1, 1),
  max.param.value = 10000)
}
\arguments{
\item{cal.cbs}{calibration period CBS. It must contain columns for frequency 
\code{x}, for recency \code{t.x} and total time observed \code{T.cal}. Optionally a 
column \code{custs} can be provided, which represents number of customers with a
specific combination of frequency \code{x}, recency \code{t.x} and \code{T.cal}.}

\item{par.start}{initial MBG/NBD parameters - a vector with \code{r}, \code{alpha}, 
\code{a} and \code{b} in that order.}

\item{max.param.value}{the upper bound on parameters}
}
\value{
list of estimated parameters
}
\description{
Estimates parameters for the MBG/NBD model via Maximum Likelihood 
Estimation.
}
\references{
Batislam, E.P., M. Denizel, A. Filiztekin. 2007. Empirical
  validation and comparison of models for customer base analysis.
  International Journal of Research in Marketing 24(3) 201-209. - Hoppe,
  Daniel, and Udo Wagner. "Customer base analysis: The case for a central
  variant of the Betageometric/NBD Model." Marketing Journal of Research and
  Management 3.2 (2007): 75-90.
}
\seealso{
\code{\link{bgcnbd.EstimateParameters}}
}

