% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbg-nbd.r
\name{mbgnbd.GenerateData}
\alias{mbgnbd.GenerateData}
\title{Simulate data according to MBG/NBD model assumptions}
\usage{
mbgnbd.GenerateData(n, T.cal, T.star, params, return.elog = FALSE)
}
\description{
Simulate data according to MBG/NBD model assumptions
}
\seealso{
\code{\link{mbgcnbd.GenerateData}}
}

