% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mcmc.r
\name{mcmc.DrawFutureTransactions}
\alias{mcmc.DrawFutureTransactions}
\title{Samples number of future transactions based on drawn parameters}
\usage{
mcmc.DrawFutureTransactions(cal.cbs, draws, T.star = cal.cbs$T.star)
}
\arguments{
\item{cal.cbs}{data.frame with column \code{t.x} and \code{T.cal}}

\item{draws}{MCMC draws returned by \code{\link{pnbd.mcmc.DrawParameters}}, \code{\link{pggg.mcmc.DrawParameters}}, \code{\link{abe.mcmc.DrawParameters}}}

\item{T.star}{length of period for which future transactions are counted}
}
\value{
2-dim array [draw x cust] with sampled future transactions
}
\description{
Samples number of future transactions based on drawn parameters
}

