% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcmc.PActive}
\alias{mcmc.PActive}
\title{Calculates P(active) based on drawn future transactions.}
\usage{
mcmc.PActive(xstar)
}
\arguments{
\item{xstar}{future transaction draws returned by \code{\link{mcmc.DrawFutureTransactions}}}
}
\value{
vector with P(active) estimates for each customer
}
\description{
Calculates P(active) based on drawn future transactions.
}

