% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.PAlive}
\alias{mcmc.PAlive}
\title{Calculates P(alive) based on MCMC parameter draws}
\usage{
mcmc.PAlive(draws)
}
\arguments{
\item{draws}{MCMC parameter draws as returned by 
\code{\link{pnbd.mcmc.DrawParameters}}, 
\code{\link{pggg.mcmc.DrawParameters}} or 
\code{\link{abe.mcmc.DrawParameters}}.}
}
\value{
Numeric vector with the customers' probabilities of being still alive
  at end of calibration period
}
\description{
Calculates P(alive) based on MCMC parameter draws
}
\examples{
cbs <- cdnow.sample()$cbs
param.draws <- pnbd.mcmc.DrawParameters(cbs, mcmc = 200, burnin = 100, thin = 20, chains = 1)
mcmc.PAlive(param.draws)
}

