% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc.plotPActiveDiagnostic}
\alias{mcmc.plotPActiveDiagnostic}
\title{Draw diagnostic plot to inspect error in P(active).}
\usage{
mcmc.plotPActiveDiagnostic(cbs, xstar,
  title = "Diagnostic Plot for P(active)")
}
\arguments{
\item{cbs}{data.frame with column \code{x} and \code{x.star}}

\item{xstar}{future transaction draws returned by \code{\link{mcmc.DrawFutureTransactions}}}

\item{title}{plot title}
}
\value{
2-element list with MCMC draws
}
\description{
Draw diagnostic plot to inspect error in P(active).
}
\examples{
cbs <- cdnow.sample()$cbs
param.draws <- pnbd.mcmc.DrawParameters(cbs, mcmc = 3000, burnin = 0, chains = 4)
xstar.draws <- mcmc.DrawFutureTransactions(cbs, param.draws)
mcmc.plotPActiveDiagnostic(cbs, xstar.draws)
}

