% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mcmc.plotPActiveDiagnostic}
\alias{mcmc.plotPActiveDiagnostic}
\title{Draw diagnostic plot to inspect error in P(active).}
\usage{
mcmc.plotPActiveDiagnostic(cbs, xstar,
  title = "Diagnostic Plot for P(active)")
}
\arguments{
\item{cbs}{data.frame with column \code{x} and \code{x.star}}

\item{xstar}{future transaction draws returned by \code{\link{mcmc.DrawFutureTransactions}}}
}
\value{
2-element list with MCMC draws
}
\description{
Draw diagnostic plot to inspect error in P(active).
}

