% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.r
\name{mcmc.setBurnin}
\alias{mcmc.setBurnin}
\title{(Re-)set burnin of MCMC chains.}
\usage{
mcmc.setBurnin(draws, burnin)
}
\arguments{
\item{draws}{MCMC draws returned by \code{\link{pnbd.mcmc.DrawParameters}}, \code{\link{pggg.mcmc.DrawParameters}}, \code{\link{abe.mcmc.DrawParameters}}}

\item{burnin}{new start index}
}
\value{
2-element list with MCMC draws
}
\description{
(Re-)set burnin of MCMC chains.
}

