% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mcmc.setBurnin}
\alias{mcmc.setBurnin}
\title{(Re-)set burnin of MCMC chains.}
\usage{
mcmc.setBurnin(draws, burnin)
}
\arguments{
\item{draws}{MCMC draws returned by \code{\link{pnbd.mcmc.DrawParameters}}, \code{\link{pcnbd.mcmc.DrawParameters}}, \code{\link{abe.mcmc.DrawParameters}}}

\item{burnin}{new start index}
}
\value{
2-element list with MCMC draws
}
\description{
(Re-)set burnin of MCMC chains.
}

