% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nbd.EstimateParameters}
\alias{nbd.EstimateParameters}
\title{Parameter Estimation for the NBD model}
\usage{
nbd.EstimateParameters(cal.cbs, par.start = c(1, 1),
  max.param.value = 10000)
}
\arguments{
\item{cal.cbs}{calibration period CBS. It must contain columns for frequency
'x' and total time observed 'T.cal'. Optionally a column 'custs' can be
provided, which represents number of customers with a specific combination
of frequency 'x' and 'T.cal'.}

\item{par.start}{initial NBD parameters - a vector with 'r' and 'alpha' in
that order.}

\item{max.param.value}{the upper bound on parameters}
}
\value{
list of estimated parameters
}
\description{
Estimates parameters for the NBD model via Maximum Likelihood Estimation.
}
\examples{
# generate artificial data
}
\references{
EHRENBERG, ASC. "The Pattern of Consumer Purchases." Quantitative
  techniques in marketing analysis: text and readings (1962): 355.
}

