% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nbd.GenerateData}
\alias{nbd.GenerateData}
\title{Simulate data according to NBD model assumptions}
\usage{
nbd.GenerateData(n, T.cal, T.star, params, return.elog = F)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{NBD parameters - a list with 'r' and 'alpha'}

\item{return.elog}{boolean - if TRUE then the event log is returned in
addition to the CBS summary}
}
\value{
list with elements 'cbs' and 'elog' containing data.frames
}
\description{
Simulate data according to NBD model assumptions
}
\seealso{
nbd.EstimateParameters
}

