% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbd.R
\name{nbd.LL}
\alias{nbd.LL}
\title{Calculate the log-likelihood of the NBD model}
\usage{
nbd.LL(params, x, T.cal)
}
\arguments{
\item{params}{NBD parameters - a vector with \code{r} and \code{alpha}, in that
order.}

\item{x}{Frequency, i.e. number of re-purchases.}

\item{T.cal}{Total time of observation period.}
}
\value{
A numeric vector of log-likelihoods.
}
\description{
Calculate the log-likelihood of the NBD model
}
\seealso{
\code{\link{nbd.cbs.LL}}
}

