% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pcnbd.GenerateData}
\alias{pcnbd.GenerateData}
\title{Generate artificial data which follows Pareto/CNBD model assumptions}
\usage{
pcnbd.GenerateData(n, T.cal, T.star, params, return.elog = F)
}
\arguments{
\item{n}{number of customers}

\item{T.cal}{length of calibration period}

\item{T.star}{length of holdout period}

\item{params}{list of parameters: {r, alpha, s, beta, t, gamma}}

\item{return.elog}{if \code{TRUE} then the event-log is returned as well; decreases performance}
}
\value{
2-elemnt list
}
\description{
Returns 2-element list
* cbs: data.frame with 'cust', \code{x}, \code{t.x}, \code{T.cal}, 'T.star', 'x.star'
       this is the summary statistics data.frame which contains all
       needed information for parameter estimation
* elog: data.frame with 'cust', \code{t}
}

