% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pcnbd.mcmc.DrawFutureTransactions}
\alias{pcnbd.mcmc.DrawFutureTransactions}
\title{Samples number of future transactions based on drawn parameters}
\usage{
pcnbd.mcmc.DrawFutureTransactions(data, draws, T.star = data$T.star)
}
\arguments{
\item{data}{data.frame with column 't.x' and 'T.cal'}

\item{draws}{MCMC draws returned by \code{pcnbd.mcmc.DrawParameters}}

\item{T.star}{length of period for which future transactions are counted}
}
\value{
2-dim array [draw x cust] with sampled future transactions
}
\description{
Samples number of future transactions based on drawn parameters
}

