% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pcnbd.mcmc.PAlive}
\alias{pcnbd.mcmc.PAlive}
\title{Calculates P(alive) based on MCMC draws}
\usage{
pcnbd.mcmc.PAlive(cal.cbs, draws)
}
\arguments{
\item{cal.cbs}{data.frame with column \code{T.cal}}

\item{draws}{MCMC draws returned by \code{\link{pcnbd.mcmc.DrawParameters}}}
}
\value{
numeric vector of probabilities
}
\description{
Calculates P(alive) based on MCMC draws
}

