% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pnbd.mcmc.DrawParameters}
\alias{pnbd.mcmc.DrawParameters}
\title{Hierarchical Bayes variant of Pareto/NBD}
\usage{
pnbd.mcmc.DrawParameters(cal.cbs, mcmc = 1500, burnin = 500, thin = 1,
  chains = 2, use_data_augmentation = TRUE, param_init = list(r = 1, alpha
  = 1, s = 1, beta = 1), hyper_prior = list(r_1 = 1/1000, r_2 = 1/1000,
  alpha_1 = 1/1000, alpha_2 = 1/1000, s_1 = 1/1000, s_2 = 1/1000, beta_1 =
  1/1000, beta_2 = 1/1000))
}
\arguments{
\item{cal.cbs}{data.frame with columns 'x', 't.x', 'T.cal'}

\item{mcmc}{number of MCMC steps}

\item{burnin}{number of initial MCMC steps which are discarded}

\item{thin}{only every thin-th MCMC step will be returned}

\item{chains}{number of MCMC chains to be run}

\item{use_data_augmentation}{determines MCMC method to be used}

\item{param_init}{list of 2nd-level parameter start values}

\item{hyper_prior}{list of hyper parameters for 2nd-level parameters}
}
\value{
2-element list:
\itemize{
 \item{\code{level_1}}{list of \code{\link{mcmc.list}} objects; one for each customer, containing individual-level draws}
 \item{\code{level_2}}{\code{\link{mcmc.list}} object containing draws of heterogeneity parameters}
}
}
\description{
\code{pnbd.mcmc.DrawParameters} samples parameters via MCMC for a given CBS
matrix
}
\details{
method 1) If \code{use_data_augmentation==TRUE} then implementation follows chapter
3.2 of Sandeep Conoor's dissertation
\url{http://gradworks.umi.com/34/02/3402149.html}, i.e. parameter space is expanded
with unobserved lifetime tau_i. Note, however, that  we follow the notation
of original SMC paper with alpha and beta being the 'rate' parameters of the
Gamma distributions, instead of 'scale' parameter.

method 2) If \code{use_data_augmentation==FALSE} then implementation follows
Shao-Hui Ma & Jin-Lan Liu paper
\url{http://ieeexplore.ieee.org/xpls/abs_all.jsp?arnumber=4344404}, i.e. no data
augmentation and draws on individual level need to be done via slice
sampling. As such it is 10x slower than method 1)

Estimating parameters via Pareto/NBD MCMC can be 10x slower than Pareto/NBD
MLE, which itself can be 10x slower than BG/NBD. Both methods exhibit highly
autocorrelated draws of {r, alpha, s, beta} and hence need to be run long, to
generate 'enough' draws
}
\examples{

# Load CDNow event log
data(cdnowElog, package="BTYD")
elog <- data.frame(t(sapply(2:nrow(cdnowElog), function(i) strsplit(as.character(cdnowElog[i,]), split=",")[[1]])))
names(elog) <- c("cust", "sampleid", "date", "cds", "sales")
elog$date <- as.Date(elog$date, "\%Y\%m\%d")

# Transform to CBS (including extra column 'litt')
cbs <- elog2cbs(elog, per="week", T.cal=as.Date("1997-09-30"))

# Pareto/NBD MCMC
set.seed(1)
pnbd.draws <- pnbd.mcmc.DrawParameters(cbs, mcmc=1000, burnin=1000, chains=2, thin=10)
summary(pnbd.draws$level_2)
plot(pnbd.draws$level_2)
summary(pnbd.draws$level_2)$quantiles[, "50\%"]

# Pareto/CNBD MCMC
set.seed(2)
param_init <- params.pnbd.mcmc
param_init$t <- 10
param_init$gamma <- 10
pcnbd.draws <- pcnbd.mcmc.DrawParameters(cbs, mcmc=100, burnin=100, chains=2, thin=2, param_init=param_init)
# Note: MCMC chain should be run longer; for demo purposes we keep runtime short here

plot(pcnbd.draws$level_2, density=FALSE)
plot(pcnbd.draws$level_2, trace=FALSE)

coda::gelman.diag(pcnbd.draws$level_2)
# -> MCMC chains have not converged yet

summary(pcnbd.draws$level_2)$quantiles[, "50\%"]

pcnbd.mcmc.plotRegularityRateHeterogeneity(pcnbd.draws)
# -> very narrow distribution around k=1; 
# CDNow customers purchases indeed seem to follow Poisson process

round(effectiveSize(pcnbd.draws$level_2))
# -> effective sample size are small for such a short chain

# draw future transaction
pnbd.xstar <- pcnbd.mcmc.DrawFutureTransactions(cbs, pnbd.draws, T.star=cbs$T.star)
pcnbd.xstar <- pcnbd.mcmc.DrawFutureTransactions(cbs, pcnbd.draws, T.star=cbs$T.star)

# calculate mean over future transaction draws for each customer
cbs$pnbd.mcmc <- apply(pnbd.xstar, 2, mean)
cbs$pcnbd.mcmc <- apply(pcnbd.xstar, 2, mean)

# calculate P(active)
cbs$pactive.pnbd.mcmc <- apply(pnbd.xstar, 2, function(x) mean(x>0))
cbs$pactive.pcnbd.mcmc <- apply(pcnbd.xstar, 2, function(x) mean(x>0))
}
\references{
Ma, Shao-Hui, and Jin-Lan Liu. "The MCMC approach for solving the Pareto/NBD model and possible extensions." Natural Computation, 2007. ICNC 2007. Third International Conference on. Vol. 2. IEEE, 2007. \url{http://ieeexplore.ieee.org/xpls/abs_all.jsp?arnumber=4344404}

Conoor, Sandeep S. Customer-base analysis in noncontractual settings. Diss. NORTHWESTERN UNIVERSITY, 2010. \url{http://gradworks.umi.com/34/02/3402149.html}
}
\seealso{
\code{\link{pcnbd.GenerateData}} \code{\link{pcnbd.DrawFutureTransactions}}
}

